% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes-helpers.R
\name{all_data_elements}
\alias{all_data_elements}
\title{Get information about all data elements on a DHIS2 server.}
\usage{
all_data_elements(
  base_url,
  username,
  password,
  path = "/api/dataElements?fields=id,name,shortName&paging=false"
)
}
\arguments{
\item{base_url}{The base URL of the DHIS2 server.}

\item{username}{The user name.}

\item{password}{The user's password.}

\item{path}{The path to the content.}
}
\description{
Get information about all data elements on a DHIS2 server.
}
