% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes-helpers.R
\name{all_datasets}
\alias{all_datasets}
\title{Get information about all the dataSets on a DHIS2 server.}
\usage{
all_datasets(
  base_url,
  username,
  password,
  path = "/api/dataSets?fields=id,name,shortName&paging=false"
)
}
\arguments{
\item{base_url}{The base URL of the DHIS2 server.}

\item{username}{The user name.}

\item{password}{The user's password.}

\item{path}{The path to the content.}
}
\value{
A \code{data.frame} with information about all the datasets
on the DHIS2 server.
}
\description{
Get information about all the dataSets on a DHIS2 server.
}
