% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes-helpers.R
\name{all_organisation_units}
\alias{all_organisation_units}
\title{Get the organisation unit identifiers and names}
\usage{
all_organisation_units(
  base_url,
  username,
  password,
  path = "/api/organisationUnits?fields=id,name,shortName&paging=false"
)
}
\arguments{
\item{base_url}{the base URL of the DHIS2 server}

\item{username}{the user name}

\item{password}{the user's password}

\item{path}{The path to the content.}
}
\description{
Get the organisation unit identifiers and names
}
