% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes.R
\name{check_dhis2_attributes}
\alias{check_dhis2_attributes}
\title{Check the DHIS2 attributes for data import}
\usage{
check_dhis2_attributes(
  username,
  password,
  base.url,
  dataset,
  organisation.unit = NULL,
  data.element.group = NULL
)
}
\arguments{
\item{username}{the user name}

\item{password}{the user's password}

\item{base.url}{the web address of the server the user wishes to log in to}

\item{dataset}{the dataSets identifiers}

\item{organisation.unit}{the organisationUnits identifiers}

\item{data.element.group}{the dataElementGroups identifiers}
}
\value{
a list
}
\description{
Check the DHIS2 attributes for data import
}
\examples{
attributes <- check_dhis2_attributes(
  username = "admin",
  password = "district",
  base.url = "https://play.dhis2.org/dev/",
  dataset = "pBOMPrpg1QX",
  organisation.unit = "DiszpKrYNg8",
  data.element.group = NULL
)
}
