% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhis2-helpers.R
\name{check_program}
\alias{check_program}
\title{Validate and retrieve program IDs}
\usage{
check_program(login, program)
}
\arguments{
\item{login}{A httr2 request object preconfigured for authentication
carrying the base url, user name, and password.}

\item{program}{A character of the program name to be validated}
}
\description{
Checks whether the specified program ID or name is valid in a DHIS2 instance.
If the name provided is a valid program name or ID, it returns
the corresponding ID; otherwise, an error message is shown and a function
listing all available programs is provided.
}
\examples{
login <- dhis2_login(base_url, user_name, password)
program_id <- check_program(
  login = login,
  program = "Malaria focus investigation"
)


}
