% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{connect_to_server}
\alias{connect_to_server}
\title{Establish connection to the server}
\usage{
connect_to_server(dbms, driver_name, host, database_name, user, password, port)
}
\arguments{
\item{dbms}{the database management system type}

\item{driver_name}{the driver name}

\item{host}{the host server name}

\item{database_name}{the database name}

\item{user}{the user name}

\item{password}{the user's password}

\item{port}{the server port ID}
}
\value{
the \code{connection} object
}
\description{
Establish connection to the server
}
\examples{
\dontrun{
con <- connect_to_server(
  dbms          = "MySQL",
  driver_name   = "",
  host          = "mysql-rfam-public.ebi.ac.uk",
  database_name = "Rfam",
  user          = "rfamro",
  password      = "",
  port          = 4497
)
}

}
