% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sormas-helpers.R
\name{convert_to_date}
\alias{convert_to_date}
\title{Convert numeric values into POSIXct,then into Date}
\usage{
convert_to_date(data, target_columns)
}
\arguments{
\item{data}{A data frame with the numeric columns to be converted into Date}

\item{target_columns}{A character vector with the name of the target numeric
columns}
}
\value{
A data frame where the specified target columns have been converted
into Date
}
\description{
Convert numeric values into POSIXct,then into Date
}
\keyword{internal}
