% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_file-helpers.R
\name{detect_separator}
\alias{detect_separator}
\title{Detect separator from a string}
\usage{
detect_separator(x)
}
\arguments{
\item{x}{a string}
}
\value{
a vector of identified separators
}
\description{
Detect separator from a string
}
\examples{
sep <- detect_separator(
 x = "My name is Karim"
)
}
