% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes.R
\name{dhis2_check_attributes}
\alias{dhis2_check_attributes}
\title{Check the DHIS2 attributes for data import}
\usage{
dhis2_check_attributes(base_url, user_name, password, query_parameters)
}
\arguments{
\item{base_url}{the web address of the server the user wishes to log in to}

\item{user_name}{the user name}

\item{password}{the user's password}

\item{query_parameters}{a list with the parameters that will be used to
determine which data is returned by the query}
}
\value{
a list of 9 elements of type \code{character}.
}
\description{
Check the DHIS2 attributes for data import
}
\examples{
\dontrun{
  attributes <- dhis2_check_attributes(
    base_url         = "https://play.dhis2.org/dev",
    user_name        = "admin",
    password         = "district",
    query_parameters = list(dataSet = "pBOMPrpg1QX",
                            orgUnit = "DiszpKrYNg8")
  )
}

}
\keyword{internal}
