% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_dhis2-helpers.R
\name{dhis2_get_attributes_from_user}
\alias{dhis2_get_attributes_from_user}
\title{Get the DHIS2 attributes from the user}
\usage{
dhis2_get_attributes_from_user(args_list)
}
\arguments{
\item{args_list}{a \code{list} of parameters provided by the user}
}
\value{
an object of type \code{list} with the values for the DHIS2 attributes.
}
\description{
Get the DHIS2 attributes from the user
}
\examples{
args <- list(dataset    = "pBOMPrpg1QX",
             start_date = "2014",
             end_date   = "2023")
arguments <- dhis2_get_attributes_from_user(args_list = args)

}
\keyword{internal}
