% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes-helpers.R
\name{dhis2_get_relevant_attributes}
\alias{dhis2_get_relevant_attributes}
\title{Get the relevant dataset}
\usage{
dhis2_get_relevant_attributes(
  base_url,
  user_name,
  password,
  attribute_id = NULL,
  which = "dataSets"
)
}
\arguments{
\item{base_url}{the web address of the server the user wishes to log in to}

\item{user_name}{the user name}

\item{password}{the user's password}

\item{attribute_id}{a vector of DHIS2 attribute ids. The ids
could be those of a dataSet or an orgUnit.}

\item{which}{the target DHIS2 end point}
}
\value{
a \code{list} of 2 elements: a \code{character} string with the target
attributes ID(s) and a \code{data.frame} that contains the data of interest
from the specified DHIS2 attribute ids.
}
\description{
Get the relevant dataset
}
\examples{
\dontrun{
result <- dhis2_get_relevant_attributes(
  base_url     = "https://play.dhis2.org/demo",
  user_name    = "admin",
  password     = "district",
  attribute_id = c("pBOMPrpg1QX", "BfMAe6Itzgt"),
  which        = "dataSets"
)
}
}
\keyword{internal}
