% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_dhis2-helpers.R
\name{dhis2_make_query_params}
\alias{dhis2_make_query_params}
\title{Make request query parameters}
\usage{
dhis2_make_query_params(dataset, org_unit, start_date = NULL, end_date = NULL)
}
\arguments{
\item{dataset}{a vector or a list of comma-separated data set identifiers}

\item{org_unit}{a vector or a list of comma-separated organisation
unit identifiers}

\item{start_date}{the start date for the time span of the values to export}

\item{end_date}{the end date for the time span of the values to export}
}
\value{
a list of the request query parameters that will be used in the API
request.
}
\description{
Make request query parameters
}
\keyword{internal}
