% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_dhis2-helpers.R
\name{dhis2_make_request_url}
\alias{dhis2_make_request_url}
\title{Make full URL from request attributes}
\usage{
dhis2_make_request_url(
  base_url,
  dataset,
  org_unit,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{base_url}{the web address of the server the user wishes to log in to}

\item{dataset}{a vector or a list of comma-separated data set identifiers}

\item{org_unit}{a vector or a list of comma-separated organisation
unit identifiers}

\item{start_date}{the start date for the time span of the values to export}

\item{end_date}{the end date for the time span of the values to export}
}
\value{
a character string with the full URL to be used in the API request
}
\description{
Make full URL from request attributes
}
\keyword{internal}
