% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_dhis2-helpers.R
\name{dhis2_subset_fields}
\alias{dhis2_subset_fields}
\title{Subset fields when reading from DHIS2}
\usage{
dhis2_subset_fields(data, fields = c("dataElement", "period", "value"))
}
\arguments{
\item{data}{the input data frame}

\item{fields}{vector of fields to select from the data frame}
}
\value{
an object of type \code{data.frame} with the data that contains only the
fields of interest.
}
\description{
Subset fields when reading from DHIS2
}
\examples{
\dontrun{
results <- dhis2_subset_fields(
  data = readepi(
    credentials_file = system.file("extdata", "test.ini",
                                     package = "readepi"),
    data_source      = "https://play.dhis2.org/demo",
    query_parameters = list(
        dataSet   = "pBOMPrpg1QX,BfMAe6Itzgt",
        orgUnit   = "DiszpKrYNg8",
        startDate = "2014",
        endDate   = "2023")
  )$data,
  fields             = c("dataElement", "period", "value")
)
}
}
\keyword{internal}
