% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_api_docs}
\alias{download_api_docs}
\title{Download the data dictionary and API specification files of a given the
health information system.}
\usage{
download_api_docs(his, path = NULL)
}
\arguments{
\item{his}{The name of the health information system of interest}

\item{path}{The path to the directory where the downloaded files will be
stored.}
}
\value{
Invisibly returns the path to the folder where the files are stored.
When \code{path = NULL}, the files will be stored in the R temporary
folder as: \code{dictionary.xlsx} and \code{api_specification.json}
respectively.
}
\description{
Download the data dictionary and API specification files of a given the
health information system.
}
\examples{
# download the SORMAS data dictionary and API specification to the temporary
# R folder
download_folder <- download_api_docs(his = "sormas")
}
