% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_file_helpers.R
\name{f}
\alias{f}
\title{Split a character by "_" and extract its element}
\usage{
f(x)
}
\arguments{
\item{x}{a character string}
}
\value{
a character string
}
\description{
Split a character by "_" and extract its element
}
\keyword{internal}
