% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{fetch_from_several_tables}
\alias{fetch_from_several_tables}
\title{Fetch data from several files}
\usage{
fetch_from_several_tables(table_names, con, fields, result)
}
\arguments{
\item{table_names}{a vector of table names}

\item{con}{the established connection to the database}

\item{fields}{a vector or a comma-separated string of column names}

\item{result}{a list of data from the table name}
}
\value{
a list of data from the tables of interest
}
\description{
Fetch data from several files
}
