% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rdbms-helpers.R
\name{fields_check}
\alias{fields_check}
\title{Check whether the user-specified fields are valid}
\usage{
fields_check(fields, table_name, login)
}
\arguments{
\item{fields}{A character vector of column names}

\item{table_name}{A character with the table name}

\item{login}{The connection object obtained from the \code{login()} function.}
}
\value{
A character vector with the valid fields
}
\description{
A valid field is the one that corresponds to a column name of the specified
table.
}
\keyword{internal}
