% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{fingertips_get_args}
\alias{fingertips_get_args}
\title{Get arguments for reading from Fingertips}
\usage{
fingertips_get_args(
  args_list = list(indicator_id = NULL, indicator_name = NULL, area_type_id = NULL,
    profile_id = NULL, profile_name = NULL, domain_id = NULL, domain_name = NULL,
    parent_area_type_id = NULL)
)
}
\arguments{
\item{args_list}{a \code{list} of user specified arguments}
}
\value{
a \code{list} of 8 elements of type \code{character} and \code{numeric} that will be
used for importing data from Fingertips
}
\description{
Get arguments for reading from Fingertips
}
\examples{
\dontrun{
args_list <- fingertips_get_args(
  list(
    indicator_id = 90362,
    area_type_id = 202
  )
)
}
}
\keyword{internal}
