% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips-helpers.R
\name{fingertips_get_id_from_dm_name}
\alias{fingertips_get_id_from_dm_name}
\title{Get indicator ID from domain name}
\usage{
fingertips_get_id_from_dm_name(
  metadata = list(indicator_profile_domain = fingertipsR::indicators(),
    indicator_ids_names = fingertipsR::indicators_unique(), area_type =
    fingertipsR::area_types()),
  domain_name = "B. Wider determinants of health",
  indicator_name = NULL
)
}
\arguments{
\item{metadata}{a list with the fingertips metadata}

\item{domain_name}{the domain name}

\item{indicator_name}{the indicator name}
}
\value{
an object of type \code{numeric} that contains the indicator ID.
}
\description{
Get indicator ID from domain name
}
\examples{
\dontrun{
indicator_id <- fingertips_get_id_from_dm_name(
  metadata    = list(
    indicator_profile_domain = fingertipsR::indicators(),
    indicator_ids_names      = fingertipsR::indicators_unique(),
    area_type                = fingertipsR::area_types()
  ),
  domain_name = "B. Wider determinants of health",
)
}
}
\keyword{internal}
