% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips-helpers.R
\name{fingertips_get_profile_name}
\alias{fingertips_get_profile_name}
\title{Get profile name from Fingertips}
\usage{
fingertips_get_profile_name(
  metadata = list(indicator_profile_domain = fingertipsR::indicators(),
    indicator_ids_names = fingertipsR::indicators_unique(), area_type =
    fingertipsR::area_types()),
  profile_id = 19L,
  profile_name = "Public Health Outcomes Framework"
)
}
\arguments{
\item{metadata}{the Fingertips metadata}

\item{profile_id}{the profile ID}

\item{profile_name}{the profile name}
}
\value{
a \code{list} of 2 elements of type \code{character} and \code{numeric}. These are
the \verb{profile name} and their correspondent indexes.
}
\description{
Get profile name from Fingertips
}
\examples{
\dontrun{
res <- fingertips_get_profile_name(
  metadata = list(
    indicator_profile_domain = fingertipsR::indicators(),
    indicator_ids_names      = fingertipsR::indicators_unique(),
    area_type                = fingertipsR::area_types()
  )
  profile_id   = 19,
  profile_name = "Public Health Outcomes Framework",
)
}
}
\keyword{internal}
