% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{fingertips_subset_rows}
\alias{fingertips_subset_rows}
\title{Subset records when reading from Fingertips}
\usage{
fingertips_subset_rows(records, id_col_name, data)
}
\arguments{
\item{records}{a vector or a comma-separated string of records}

\item{id_col_name}{the column name with the subject IDs}

\item{data}{the data read from Fingertips}
}
\value{
a data frame with the records of interest
}
\description{
Subset records when reading from Fingertips
}
