% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_genomic_data.R
\name{genome_assembly}
\alias{genome_assembly}
\title{Function to perform the genome assembly}
\usage{
genome_assembly(
  input.dir,
  assembly.dir,
  consensus.fasta.dir,
  num.samples,
  device = "MinION",
  partition = NULL
)
}
\arguments{
\item{input.dir}{the path to the folder with the input fastq files. Each sample/barcode will have a directory in this folder}

\item{assembly.dir}{the path to the folder where the output files from the genome assembly will be stored}

\item{consensus.fasta.dir}{the path to the folder where the merged consensus fasta file will be stored}

\item{num.samples}{the number of samples in the dataset}

\item{device}{the sequencing device (either MinION or PromethION)}

\item{partition}{the partition. Only set this when running on a HPC, default is NULL}
}
\value{
the output of this function is the merged fasta file that will be in the folder specified using the consensus.fasta.dir parameter
}
\description{
Function to perform the genome assembly
}
