% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes-helpers.R
\name{get_data_elements}
\alias{get_data_elements}
\title{Get one or more data element groups.}
\usage{
get_data_elements(data_element_group_ids, base_url, username, password)
}
\arguments{
\item{data_element_group_ids}{The dataElementGroups identifiers.}

\item{base_url}{The base URL of the DHIS2 server.}

\item{username}{The user name.}

\item{password}{The user's password.}
}
\value{
A \code{list} with the data elements of interest.
}
\description{
Get one or more data element groups.
}
