% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes-helpers.R
\name{get_datasets}
\alias{get_datasets}
\title{Get one or more datasets from the DHIS2 server.}
\usage{
get_datasets(dataset_ids, base_url, username, password)
}
\arguments{
\item{dataset_ids}{The dataSets identifiers.}

\item{base_url}{The base URL of the DHIS2 server.}

\item{username}{The user name.}

\item{password}{The user's password.}
}
\value{
A \code{list} with the relevant datasets.
}
\description{
Get one or more datasets from the DHIS2 server.
}
\examples{
result <- get_datasets(
  dataset_ids = "pBOMPrpg1QX,BfMAe6Itzgt",
  base_url = "https://play.dhis2.org/dev/",
  username = "admin",
  password = "district"
)
}
