% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default_filters}
\alias{get_default_filters}
\title{Return the default list of data filtration options.}
\usage{
get_default_filters()
}
\value{
A list with the pre-defined data filtration options.
}
\description{
Return the default list of data filtration options.
}
\keyword{internal}
