% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes_helpers.R
\name{get_dhis2_attributes}
\alias{get_dhis2_attributes}
\title{Get the target DHIS2 attribute identifiers and names}
\usage{
get_dhis2_attributes(
  base_url = "https://play.dhis2.org/dev/",
  username = "admin",
  password = "district",
  which = "dataSets"
)
}
\arguments{
\item{base_url}{the base URL of the DHIS2 server}

\item{username}{the user name}

\item{password}{the user's password}

\item{which}{the target DHIS2 attribute name.}
}
\value{
an object of type \code{data.frame} with details about the DHIS2
attributes of interest.
}
\description{
Get the target DHIS2 attribute identifiers and names
}
\examples{
\dontrun{
datasets <- get_dhis2_attributes(
  base_url = "https://play.dhis2.org/dev/",
  username = "admin",
  password = "district",
  which = "dataSets"
)
}
}
