% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_disease_names}
\alias{get_disease_names}
\title{Get the list of disease names for which cases data is available in a
given health information system (HIS).}
\usage{
get_disease_names(from, user_name, password)
}
\arguments{
\item{from}{The name of the HIS. Possible values are: sormas, dhis2, odk,
redcap, global.health, goData}

\item{user_name}{The user name}

\item{password}{The user's password}
}
\value{
A vector of the list of disease names.
}
\description{
Get the list of disease names for which cases data is available in a
given health information system (HIS).
}
\examples{
# get the disease names from SORMAS
sormas_diseases <- get_disease_names(
  from = "sormas",
  user_name = "SurvSup",
  password = "Lk5R7JXeZSEc"
)

}
