% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_endpoints}
\alias{get_endpoints}
\title{Get the list of available endpoints from an API from the API specification
documentation.}
\usage{
get_endpoints(his)
}
\arguments{
\item{his}{A character that represents the name of the health information
system.}
}
\value{
A vector of the endpoints found from the API specification document.
}
\description{
Get the list of available endpoints from an API from the API specification
documentation.
}
\examples{
# get the list of endpoints from SORMAS
sormas_endpoints <- get_endpoints(his = "sormas")

}
