% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhis2-helpers.R
\name{get_entity_attributes}
\alias{get_entity_attributes}
\title{Get the attribute names and ids of the tracked entities}
\usage{
get_entity_attributes(x)
}
\arguments{
\item{x}{A list with the tracked entity attributes}
}
\value{
A data frame with two columns representing the tracked entity
attributes IDs and display names.
}
\description{
Use this function to get the ids and names of the features collected about
each tracked entities. The ids and names will be used to update the list
of query parameters.
}
\keyword{internal}
