% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhis2-helpers.R
\name{get_entity_data}
\alias{get_entity_data}
\title{Get tracked entity attributes}
\usage{
get_entity_data(login, api_version, tracked_entities, org_units)
}
\arguments{
\item{login}{A httr2_response object returned by the \code{dhis2_login()} function}

\item{api_version}{A numeric with the API version obtained from the
\code{get_api_version()} function}

\item{org_units}{A data frame with all organisation units from target DHIS2
instance. This is the output from the \code{get_organisation_units()} function}
}
\value{
A data frame with the tracked entity attributes
}
\description{
Get tracked entity attributes
}
\keyword{internal}
