% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhis2-helpers.R
\name{get_event_ids}
\alias{get_event_ids}
\title{Get event IDs for specific tracked entity}
\usage{
get_event_ids(tracked_entity)
}
\arguments{
\item{tracked_entity}{A list with tracked entity data}
}
\value{
A data frame of event IDs associated with the specific tracked
entity
}
\description{
Get event IDs for specific tracked entity
}
\keyword{internal}
