% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_extension}
\alias{get_extension}
\title{Get file extension}
\usage{
get_extension(file_path)
}
\arguments{
\item{file_path}{the target file path}
}
\value{
a string that corresponds to the file extension
}
\description{
Get file extension
}
\examples{
ext <- get_extension(file_path = system.file("extdata", "test.txt",
package = "readepi"))
}
