% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fields}
\alias{get_fields}
\title{Get the field names from an endpoint}
\usage{
get_fields(his, endpoint)
}
\arguments{
\item{his}{A character with the name of the HIS}

\item{endpoint}{A character with the name of the target endpoint from which
you want to get the field names.}
}
\value{
A vector of characters representing all the fields found within the
specified endpoint.
}
\description{
Get the field names from an endpoint
}
\examples{
# get all field from the 'persons' endpoint in 'SORMAS'
fields <- get_fields(
  his = "sormas",
  endpoint = "persons"
)
}
