% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_file_helpers.R
\name{get_file_name}
\alias{get_file_name}
\title{Get the file base name}
\usage{
get_file_name(result, idx, base_name)
}
\arguments{
\item{result}{a \code{list} of \code{data.frame}}

\item{idx}{an integer}

\item{base_name}{a \code{character} with the file base name}
}
\value{
a \code{character} that will be used to name the file.
}
\description{
Every data frame in the output list is named after its correspondent file
base name. This function will extract that base name appropriately and add
an index to it when several files have the same name.
}
\keyword{internal}
