% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_filters}
\alias{get_filters}
\title{Construct the list of parameters used to subset rows for the imported data
from the user-provided filters.}
\usage{
get_filters(his, filter)
}
\arguments{
\item{his}{A character with the name of the target HIS}

\item{filter}{A list with the filters to apply on the imported data}
}
\value{
An updated list of filters with the user-specified parameters.
}
\description{
Construct the list of parameters used to subset rows for the imported data
from the user-provided filters.
}
\keyword{internal}
