% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips_helpers.R
\name{get_ind_id_from_domain_name}
\alias{get_ind_id_from_domain_name}
\title{Get indicator ID from domain name}
\usage{
get_ind_id_from_domain_name(metadata, domain_name, indicator_name = NULL)
}
\arguments{
\item{metadata}{a list with the fingertips metadata}

\item{domain_name}{the domain name}

\item{indicator_name}{the indicator name}
}
\value{
an object of type \code{numeric} that contains the indicator ID
}
\description{
Get indicator ID from domain name
}
\examples{
\dontrun{
indicator_id <- get_ind_id_from_domain_name(
  metadata = list(
    indicator_profile_domain = fingertipsR::indicators(),
    indicator_ids_names = fingertipsR::indicators_unique(),
    area_type = fingertipsR::area_types()
  ),
  domain_name = "B. Wider determinants of health",
)
}
}
\keyword{internal}
