% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips-helpers.R
\name{get_ind_id_from_profile}
\alias{get_ind_id_from_profile}
\title{Get indicator ID from profile ID and/or profile name}
\usage{
get_ind_id_from_profile(
  metadata,
  domain_id = NULL,
  domain_name = NULL,
  indicator_name = NULL,
  profile_name = NULL,
  profile_id = NULL
)
}
\arguments{
\item{metadata}{a list with the fingertips metadata}

\item{domain_id}{the domain ID}

\item{domain_name}{the domain name}

\item{indicator_name}{the indicator name}

\item{profile_name}{the profile name}

\item{profile_id}{the profile ID}
}
\value{
an object of type \code{numeric} that contains the indicator ID
}
\description{
Get indicator ID from profile ID and/or profile name
}
\examples{
\dontrun{
  res <- get_ind_id_from_profile(
    metadata = list(
      indicator_profile_domain = fingertipsR::indicators(),
      indicator_ids_names = fingertipsR::indicators_unique(),
      area_type = fingertipsR::area_types()
    ),
    profile_id = 19
  )
}

}
