% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_indicatorID_from_domainID}
\alias{get_indicatorID_from_domainID}
\title{Get indicator ID from domain ID}
\usage{
get_indicatorID_from_domainID(metadata, domain_id, indicator_name = NULL)
}
\arguments{
\item{metadata}{a list with the fingertips metadata}

\item{domain_id}{the domain ID}

\item{indicator_name}{the indicator name}
}
\value{
the indicator ID
}
\description{
Get indicator ID from domain ID
}
