% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_indicatorID_from_domainName}
\alias{get_indicatorID_from_domainName}
\title{Get indicator ID from domain name}
\usage{
get_indicatorID_from_domainName(metadata, domain_name, indicator_name = NULL)
}
\arguments{
\item{metadata}{a list with the fingertips metadata}

\item{domain_name}{the domain name}

\item{indicator_name}{the indicator name}
}
\value{
the indicator ID
}
\description{
Get indicator ID from domain name
}
