% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_indicatorID_from_indicatorName}
\alias{get_indicatorID_from_indicatorName}
\title{Get indicator ID from indicator name}
\usage{
get_indicatorID_from_indicatorName(metadata, indicator_name)
}
\arguments{
\item{metadata}{a list with the fingertips metadata}

\item{indicator_name}{the indicator name}
}
\value{
the indicator ID
}
\description{
Get indicator ID from indicator name
}
