% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhis2-helpers.R
\name{get_org_unit_as_long}
\alias{get_org_unit_as_long}
\title{Transform the organisation units data frame from large to long format}
\usage{
get_org_unit_as_long(login, org_units = NULL)
}
\arguments{
\item{login}{The login object}

\item{org_units}{A data frame of all the organisation units obtained from the
\code{get_organisation_units()} function. Default is \code{NULL}.}
}
\value{
A data with three columns corresponding to the organisation unit
names, IDs, and levels
}
\description{
Transform the organisation units data frame from large to long format
}
\keyword{internal}
