% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhis2-helpers.R
\name{get_org_unit_levels}
\alias{get_org_unit_levels}
\title{Extract the DHSI2 organization unit's hierarchical levels.}
\usage{
get_org_unit_levels(login)
}
\arguments{
\item{login}{A httr2_response object returned by the \code{dhis2_login()} function}
}
\value{
A data frame with three columns containing the organization unit's
names, IDs, and their hierarchical levels.
}
\description{
The level is a numerical number, with 1 referring to the "Country",
2 "Region", and so on to the deepest level denoting the health care
reporting unit.
}
\keyword{internal}
