% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhis2-helpers.R
\name{get_org_units}
\alias{get_org_units}
\title{Retrieve  DHIS2 organization units, along with their IDs, names,
parent IDs, and levels.}
\usage{
get_org_units(login)
}
\arguments{
\item{login}{A httr2_response object returned by the \code{dhis2_login()} function}
}
\value{
A data frame of organization units with the following fields:
organisation unit id and name, as well as the corresponding parent
organisation unit id, and level.
}
\description{
Retrieve  DHIS2 organization units, along with their IDs, names,
parent IDs, and levels.
}
\keyword{internal}
