% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_read_file_args}
\alias{get_read_file_args}
\title{Get arguments for reading from files}
\usage{
get_read_file_args(args_list)
}
\arguments{
\item{args_list}{a list of user specified arguments}
}
\value{
a list of the parameters to be used for reading from file
}
\description{
Get arguments for reading from files
}
