% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_read_fingertips_args}
\alias{get_read_fingertips_args}
\title{Get arguments for reading from Fingertips}
\usage{
get_read_fingertips_args(args_list)
}
\arguments{
<<<<<<< HEAD
\item{args_list}{a \code{list} of user specified arguments}
}
\value{
a \code{list} of 8 elements of type \code{character} or \code{numeric} that will be
used for importing data from Fingertips
=======
\item{args_list}{a list of user specified arguments}
}
\value{
a list of the parameters to be used for reading from Fingertips
>>>>>>> main
}
\description{
Get arguments for reading from Fingertips
}
\examples{
<<<<<<< HEAD
\dontrun{
args_list <- get_read_fingertips_args(
  list(
    indicator_id = 90362,
    area_type_id = 202
  )
)
}
}
\keyword{internal}
=======
args_list <- get_read_fingertips_args(
 list(
   indicator_id = 90362,
   area_type_id = 202
 )
)
}
>>>>>>> main
