% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes-helpers.R
\name{get_relevant_data_elt_group}
\alias{get_relevant_data_elt_group}
\title{Get the relevant data element groups}
\usage{
get_relevant_data_elt_group(data_element_group, base_url, username, password)
}
\arguments{
\item{data_element_group}{the dataElementGroups identifiers}

\item{base_url}{the web address of the server the user wishes to log in to}

\item{username}{the user name}

\item{password}{the user's password}
}
\value{
a \code{list} of 2 elements: a \code{character} string with the target
data element group ID(s) and a \code{data.frame} of the details about all
data element group in the target DHIS2 server.
}
\description{
Get the relevant data element groups
}
