% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_relevant_dataset}
\alias{get_relevant_dataset}
\title{Get the relevant dataset}
\usage{
get_relevant_dataset(dataset, base_url, username, password)
}
\arguments{
\item{dataset}{the dataSets identifiers}

\item{base_url}{the web address of the server the user wishes to log in to}

\item{username}{the user name}

\item{password}{the user's password}
}
\value{
a list with the relevant datasets
}
\description{
Get the relevant dataset
}
\examples{
result <- get_relevant_dataset(
dataset = "pBOMPrpg1QX,BfMAe6Itzgt",
base_url = "https://play.dhis2.org/dev/",
username = "admin",
password = "district"
)
}
