% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_relevant_organisation_unit}
\alias{get_relevant_organisation_unit}
\title{Get the relevant organisation units}
\usage{
get_relevant_organisation_unit(organisation_unit, base_url, username, password)
}
\arguments{
\item{organisation_unit}{the organisationUnits identifiers}

\item{base_url}{the web address of the server the user wishes to log in to}

\item{username}{the user name}

\item{password}{the user's password}
}
\value{
a list with the relevant organisation units
}
\description{
Get the relevant organisation units
}
