% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhis2-helpers.R
\name{get_request_params}
\alias{get_request_params}
\title{Get the request query parameters for a specific API version}
\usage{
get_request_params(api_version)
}
\arguments{
\item{api_version}{A numeric with the API version}
}
\value{
A data frame with single row containing the query parameters needed
to submit a request
}
\description{
Get the request query parameters for a specific API version
}
\keyword{internal}
