% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{getExtension}
\alias{getExtension}
\title{Get file extension}
\usage{
getExtension(file.path)
}
\arguments{
\item{file.path}{the target file path}
}
\value{
a string that corresponds to the file extension
}
\description{
Get file extension
}
\examples{
ext = getExtension(file.path=system.file("extdata", "test.txt", package = "readepi"))
}
