% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{identify_tables_and_queries}
\alias{identify_tables_and_queries}
\title{Identify queries or tables names from the user supplied string}
\usage{
identify_tables_and_queries(src, tables)
}
\arguments{
\item{src}{the user supplied string. This is usually an SQL query or a table
name or a combination of both.}

\item{tables}{the list of all tables in the database}
}
\value{
a list with the identified queries and tables
}
\description{
Identify queries or tables names from the user supplied string
}
\examples{
test <- identify_tables_and_queries(
  src    = "select * from author",
  tables = c("author", "karim")
)
}
\keyword{internal}
