% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_file_helpers.R
\name{import_non_rio_files}
\alias{import_non_rio_files}
\title{Import files with extensions that are not handled by the {rio} package}
\usage{
import_non_rio_files(
  files,
  files_base_names,
  files_extensions,
  result = list()
)
}
\arguments{
\item{files}{a vector of the file paths}

\item{files_base_names}{a \code{vector} of the file base names}

\item{files_extensions}{a \code{vector} of the file extension}

\item{result}{a \code{list} object where the data read from each file will be
stored as a \code{data.frame}}
}
\value{
an object of type \code{list} with 1 or several object of type
\code{data.frame}. Each elements of the list is named after the file from
which the corresponding data is coming.
}
\description{
Import files with extensions that are not handled by the {rio} package
}
\keyword{internal}
