% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{install_driver}
\alias{install_driver}
\title{install MS SQL ODBC driver}
\usage{
install_driver(driver_version, force_install)
}
\arguments{
\item{driver_version}{the MS ODBC driver version of interest}

\item{force_install}{if TRUE, the selected MS ODBC driver version will be force installed}
}
\description{
install MS SQL ODBC driver
}
