% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{install_odbc_driver_linux}
\alias{install_odbc_driver_linux}
\title{install ODBC driver for Linux}
\usage{
install_odbc_driver_linux(driver_version, force_install)
}
\description{
install ODBC driver for Linux
}
