% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{install_odbc_driver_mac}
\alias{install_odbc_driver_mac}
\title{install ODBC driver for Mac}
\usage{
install_odbc_driver_mac(driver_version, force_install)
}
\description{
install ODBC driver for Mac
}
