% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{install_odbc_driver_ubuntu}
\alias{install_odbc_driver_ubuntu}
\title{Install MS SQL ODBC driver for Ubuntu}
\usage{
install_odbc_driver_ubuntu(driver_version, force_install)
}
\arguments{
\item{driver_version}{the MS ODBC driver version of interest}

\item{force_install}{if TRUE, the selected MS ODBC driver version will be force installed}
}
\description{
Install MS SQL ODBC driver for Ubuntu
}
