% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lookup_table.R
\docType{data}
\name{lookup_table}
\alias{lookup_table}
\title{Lookup table}
\format{
A data frame of 11 rows and 3 columns
}
\usage{
lookup_table
}
\description{
The lookup table is used to establish the correspondence between R operators,
provided in an expression, and the DHIS2 or SQL operators.
}
\keyword{datasets}
