% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_default_filters}
\alias{modify_default_filters}
\title{Update the data filtration options with the user-specified values.}
\usage{
modify_default_filters(defaults, filters, strict = TRUE)
}
\arguments{
\item{defaults}{A list with the default data filtration options}

\item{filters}{A list with the data filtration options defined by the user}

\item{strict}{A boolean used to specify whether to not allow for other data
filtration options than the defaults. Default is \code{TRUE}.}
}
\value{
A list with the updated data filtration options.
}
\description{
Update the data filtration options with the user-specified values.
}
\keyword{internal}
