% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rdbms-helpers.R
\name{query_check}
\alias{query_check}
\title{Check whether the user-provided query is valid}
\usage{
query_check(query, login, table_name)
}
\arguments{
\item{query}{An R expression that will be converted into an SQL query}

\item{login}{The connection object obtained from the \code{login()} function.}

\item{table_name}{A character with the table name}
}
\value{
Invisibly returns \code{TRUE} if the query is valid; throws an error
otherwise.
}
\description{
We define a query as valid when it contains either one of the column names of
the table being queried and/or one of the \code{R} operators provided in the
\code{lookup_table} object of the package.
}
\keyword{internal}
