% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_file-helpers.R
\name{read_files}
\alias{read_files}
\title{Sub-function for reading files using rio package}
\usage{
read_files(sep, file_path, which, format)
}
\arguments{
\item{sep}{the separator between the columns in the file}

\item{file_path}{the path to the file to be read}

\item{which}{a string used to specify the name of the excel sheet to import}

\item{format}{a string used to specify the file format}
}
\value{
a list of data frames where each contains data from a file
}
\description{
Sub-function for reading files using rio package
}
