% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{read_files_in_directory}
\alias{read_files_in_directory}
\title{Sub-function for reading file in a directory}
\usage{
read_files_in_directory(file_path, pattern)
}
\arguments{
\item{file_path}{the path to the file to be read}

\item{pattern}{when specified, only files with this suffix will be imported}
}
\value{
a list of data frames where each contains data read from a file
}
\description{
Sub-function for reading file in a directory
}
