% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stuff.R
\name{read_from_file}
\alias{read_from_file}
\title{function to read data from all supported file types in the \code{rio} package.}
\usage{
read_from_file(
  file.path,
  sep = NULL,
  format = NULL,
  which = NULL,
  pattern = NULL
)
}
\arguments{
\item{file.path}{the path to the file to be read. When several files need to be imported from a directory, this should be the path to that directory}

\item{sep}{the separator between the columns in the file. This is only required for space-separated files.}

\item{format}{a string used to specify the file format. This is useful when a file does not have an extension, or has a file extension that does not match its actual type}

\item{which}{a string used to specify the name of the excel sheet to import}

\item{pattern}{when specified, only files with this suffix will be imported from the specified directory}
}
\value{
a data frame (import from file) or a list (import several files from directory)
}
\description{
function to read data from all supported file types in the \code{rio} package.
}
\examples{
data = read_from_file(file.path = "/path/to/test.txt", sep=NULL, format=NULL, which=NULL, pattern=NULL)
}
