% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips.R
\name{read_from_fingertips}
\alias{read_from_fingertips}
\title{Read from Fingertips}
\usage{
read_from_fingertips(
  indicator_id = NULL,
  indicator_name = NULL,
  area_type_id = NULL,
  parent_area_type_id = NULL,
  profile_id = NULL,
  profile_name = NULL,
  domain_id = NULL,
  domain_name = NULL,
  fields = NULL,
  records = NULL,
  id_position = NULL,
  id_col_name = NULL
)
}
\arguments{
\item{indicator_id}{a numeric vector of indicator IDs}

\item{indicator_name}{a vector of a comma-separated list of indicator names}

\item{area_type_id}{a vector of area type IDs}

\item{parent_area_type_id}{a vector of parent area type IDs}

\item{profile_id}{a vector of profile IDs}

\item{profile_name}{a vector or a comma-separated list of profile names}

\item{domain_id}{a vector of domain IDs}

\item{domain_name}{a vector or a comma-separated list of domain names}

\item{fields}{a vector or a comma-separated string of column names.
If provided, only those columns will be imported.}

\item{records}{a vector or a comma-separated string of records.
When specified, only these records will be imported.}

\item{id_position}{the column position of the variable that unique identifies
the subjects. When the name of the column with the subject IDs is known,
this can be provided using the \code{id_col_name} argument}

\item{id_col_name}{the column name with the subject IDs.}
}
\value{
a \code{list} of 1 element of type \code{data.frame}. This contains the
imported dataset of interest
}
\description{
Read from Fingertips
}
\examples{
\dontrun{
data <- read_from_fingertips(indicator_id = 90362, area_type_id = 202)
}
}
\keyword{internal}
