% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stuff.R
\name{read_from_ms_sql_server}
\alias{read_from_ms_sql_server}
\title{function to read data from relational databases hosted by online MS SQL server. The user needs to have read access to the database et should install the appropriate MS driver that is compatible with the SQLServer version}
\usage{
read_from_ms_sql_server(
  user,
  password,
  host,
  port,
  database.name,
  driver.name,
  table.names = NULL,
  records = NULL,
  fields = NULL,
  id.position = 1
)
}
\arguments{
\item{user}{the user name}

\item{password}{the user password}

\item{host}{the name of the host server}

\item{port}{the port ID}

\item{database.name}{the name of the database that contains the table from which the data should be pulled}

\item{driver.name}{the name of the MS driver. use \code{odbc::odbcListDrivers()} to display the list of installed drivers}

\item{table.names}{a vector or a comma-separated list of table names from the project or database. When this is not specified, the function will extract data from all tables in the database.}

\item{records}{a vector or a comma-separated string of subset of subject IDs. When specified, only the records that correspond to these subjects will be imported.}

\item{fields}{a vector of strings where each string is a comma-separated list of column names. The element of this vector should be a list of column names from the first table specified in the \code{table.names} argument and so on...}

\item{id.position}{a vector of the column positions of the variable that unique identifies the subjects in each table. This should only be specified when the column with the subject IDs is not the first column. default is 1.}
}
\value{
a data frame
}
\description{
function to read data from relational databases hosted by online MS SQL server. The user needs to have read access to the database et should install the appropriate MS driver that is compatible with the SQLServer version
}
\examples{
\dontrun{
data <- read_from_ms_sql_server(user = "kmane", password = "****", host = "robin.mrc.gm", port = 1433, database.name = "my_db", table.names = "my_table", driver.name = "ODBC Driver 17 for SQL Server")
}
}
