% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{read_multiple_files}
\alias{read_multiple_files}
\title{Read multiple files, including multiple files in a directory}
\usage{
read_multiple_files(files, dirs, format = NULL, which = NULL)
}
\description{
Read multiple files, including multiple files in a directory
}
