% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{read_rio_formats}
\alias{read_rio_formats}
\title{Title}
\usage{
read_rio_formats(files_extensions, rio_extensions, files, files_base_names)
}
\arguments{
\item{files_extensions}{a vector a file extensions made from your files of
interest}

\item{rio_extensions}{a vector of files extensions supported by the rio
package}

\item{files}{a vector a files of interest}

\item{files_base_names}{a vector of file base}
}
\value{
a list a the parameters needed to import data using rio
}
\description{
Title
}
