% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_servers.R
\name{read_server}
\alias{read_server}
\title{Read data from database management systems (DBMS).}
\usage{
read_server(connexion, query, endpoint = NULL)
}
\arguments{
\item{name}{the name of the database management system}

\item{credentials}{an object of class credentials}

\item{server}{an object of class server}

\item{filter}{a \code{list} that is used to specify the data source and filtration
conditions. It should contain the following elements:
\enumerate{
\item src: the name of the target table(s) or an SQL query
\item fields: a vector of the name of the columns that will be used to
subset the data. Default is \code{NULL}.
\item values: a vector with the values of interest from the column
specified specified in the fields argument.
This should be of the same length as 'fields'. Default is \code{NULL}.
\item target_columns: a vector of column names. When specified, only
those columns will be returned. Default is "all".
}}
}
\value{
a \code{list} of 1 or several objects of type \code{data.frame}. The number of
elements in the list depends on the number of tables from which the
data is fetched.
}
\description{
The function assumes the user has read access to the database.
Importing data from DBMS requires the installation of the appropriate
\code{driver} that is compatible with the server version hosting the database.
For more details, see the \href{../vignettes/install_drivers.Rmd}{vignette} on
how to install the driver.
}
\examples{
\dontrun{
  # establish the connection to the database
  conn <- authenticate(
    name          = "MySQL",
    base_url      = "mysql-rfam-public.ebi.ac.uk",
    user_name     = "rfamro",
    password      = "",
    driver_name   = "",
    database_name = "Rfam",
    port          = 4497
  )

  # import data from the 'author' table
  data <- read_from_server(
    connexion = conn,
    src       = "author",
    filter    = NULL,
    select    = NULL
  )
}
}
