% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stuff.R
\name{readepi}
\alias{readepi}
\title{Function to import epidemiology-related data from different sources into R}
\usage{
readepi(
  credentials.file = NULL,
  file.path = NULL,
  sep = NULL,
  format = NULL,
  which = NULL,
  pattern = NULL,
  project.id = NULL,
  driver.name = NULL,
  table.name = NULL,
  records = NULL,
  fields = NULL,
  id.position = 1
)
}
\arguments{
\item{credentials.file}{the path to the file with the user-specific credential details for the projects of interest. It is required when importing data from DBMS. This is a tab-delimited file with the following columns:
\enumerate{
\item user_name: the user name
\item password: the user password (for REDCap, this corresponds to the \strong{token} that serves as password to the project)
\item host_name: the host name (for MS SQL servers) or the URI (for REDCap)
\item project_id: the project ID or the name of the database you are access to.
\item comment: a summary description about the project or database of interest
\item dbms: the name of the DBMS: 'redcap' or 'REDCap' when reading from REDCap, 'sqlserver' or 'SQLServer' when reading from MS SQLServer
\item port: the port ID (used for MS SQLServers)
}}

\item{file.path}{the path to the file to be read. When several files need to be imported from a directory, this should be the path to that directory}

\item{sep}{the separator between the columns in the file. This is only required for space-separated files.}

\item{format}{a string used to specify the file format. This is useful when a file does not have an extension, or has a file extension that does not match its actual type}

\item{which}{a string used to specify which objects should be extracted (the name of the excel sheet to import)}

\item{pattern}{when specified, only files with this suffix will be imported from the specified directory}

\item{project.id}{for relational DB, this is the name of the database that contains the table from which the data should be pulled. Otherwise, it is the project ID you were given access to.}

\item{driver.name}{the name of the MS driver. use \code{odbc::odbcListDrivers()} to display the list of installed drivers}

\item{table.name}{the name of the target table}

\item{records}{a vector or a comma-separated string of subset of subject IDs. When specified, only the records that correspond to these subjects will be imported.}

\item{fields}{a vector or a comma-separated string of column names. If provided, only those columns will be imported.}

\item{id.position}{the column position of the variable that unique identifies the subjects. This should only be specified when the column with the subject IDs is not the first column. default is 1.}
}
\value{
a list with 2 data frames (data and metadata) when reading from REDCap. A data frame otherwise.
}
\description{
the function allows import of data stored in database management systems (DBMS) including both relational and NoSQL databases. It also contains functions to import data from common file types such as csv, txt, xlsx, xml, json, etc.
}
\examples{
data <- readepi(credentials.file = system.file("extdata", "test.ini", package = "readepi"), project.id = "Pats__Covid_19_Cohort_1_Screening")
}
