% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stuff.R
\name{readMSsqlCredentials}
\alias{readMSsqlCredentials}
\title{function to read user credentials from file, deprecated}
\usage{
readMSsqlCredentials(credentials.file)
}
\arguments{
\item{credentials.file}{the file with the user credentials}
}
\value{
a list with the user credentials details
}
\description{
function to read user credentials from file, deprecated
}
