% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{redcap_get_results}
\alias{redcap_get_results}
\title{Check and return data fetch from redcap}
\usage{
redcap_get_results(redcap_data, metadata)
}
\arguments{
\item{redcap_data}{the object with redcap data}

\item{metadata}{the object with redcap metadata}
}
\value{
a list with the redcap dataset and its associated metadata as
data frames
}
\description{
Check and return data fetch from redcap
}
