% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_redcap-helpers.R
\name{redcap_import_data}
\alias{redcap_import_data}
\title{Import data from REDCap under all scenari}
\usage{
redcap_import_data(
  base_url,
  token,
  records = NULL,
  fields = NULL,
  id_position = 1L,
  id_col_name = NULL
)
}
\arguments{
\item{base_url}{the URI of the server}

\item{token}{the user-specific string that serves as the password for a
project}

\item{records}{a vector or a comma-separated string of subset of subject IDs}

\item{fields}{a vector or a comma-separated string of column names}

\item{id_position}{the column position of the variable that unique identifies
the subjects}

\item{id_col_name}{the column name with the subject IDs}
}
\value{
a \code{list} of 2 elements of type \code{data.frame}. These are the dataset of
interest and its associated metadata.
}
\description{
This is a wrapper across all the use case of reading data from
REDCap i.e. around the function that all records and fields from the
project, around the function that read specific records/fields or both at
the same time
}
\examples{
\dontrun{
result <- redcap_import_data(
  base_url    = "https://bbmc.ouhsc.edu/redcap/api/",
  token       = "9A81268476645C4E5F03428B8AC3AA7B",
  records     = c("1", "3", "5"),
  fields      = c("record_id", "name_first", "age", "bmi"),
  id_position = 1
  id_col_name = NULL
)
}
}
\keyword{internal}
