% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_redcap-helpers.R
\name{redcap_read}
\alias{redcap_read}
\title{Read all rows and columns from redcap}
\usage{
redcap_read(uri, token, id_position)
}
\arguments{
\item{uri}{the URI of the REDCap project}

\item{token}{the user-specific string that serves as the password for a
project}

\item{id_position}{the column position of the variable that unique identifies
the subjects}
}
\value{
a \code{list} of 2 elements of type \code{data.frame} that contain the project
data and its associated metadata.
}
\description{
Read all rows and columns from redcap
}
