% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_redcap-helpers.R
\name{redcap_read_data}
\alias{redcap_read_data}
\title{Read all rows and columns from redcap}
\usage{
redcap_read_data(base_url, token, id_position = 1L)
}
\arguments{
\item{base_url}{the URI of the REDCap project}

\item{token}{the user-specific string that serves as the password for a
project}

\item{id_position}{the column position of the variable that unique identifies
the subjects}
}
\value{
a \code{list} of 2 elements of type \code{data.frame} that contain the project
data and its associated metadata.
}
\description{
Read all rows and columns from redcap
}
\examples{
\dontrun{
  result <- redcap_read_data(
    base_url    = "https://bbmc.ouhsc.edu/redcap/api/",
    token       = "9A81268476645C4E5F03428B8AC3AA7B",
    id_position = 1L
  )
}

}
\keyword{internal}
