% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_redcap-helpers.R
\name{redcap_read_records}
\alias{redcap_read_records}
\title{Subset records from a REDCap project}
\usage{
redcap_read_records(
  base_url,
  token,
  records,
  id_position = 1L,
  id_col_name = NULL
)
}
\arguments{
\item{base_url}{the URI of the REDCap project}

\item{token}{the user-specific string that serves as the password for a
project}

\item{records}{a vector or a comma-separated string of subset of subject IDs}

\item{id_position}{the column position of the variable that unique identifies
the subjects}

\item{id_col_name}{the column name with the subject IDs}
}
\value{
a \code{list} of 2 elements of type \code{data.frame} that contain the project
data with the records of interest and its associated metadata.
}
\description{
Subset records from a REDCap project
}
\keyword{internal}
