% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_redcap-helpers.R
\name{redcap_read_rows_columns}
\alias{redcap_read_rows_columns}
\title{Subset records and columns from a REDCap project}
\usage{
redcap_read_rows_columns(
  base_url,
  token,
  fields = NULL,
  records = NULL,
  id_position = 1L,
  id_col_name = NULL
)
}
\arguments{
\item{base_url}{the URI of the REDCap project}

\item{token}{the user-specific string that serves as the password for a
project}

\item{fields}{a vector or a comma-separated string of column names}

\item{records}{a vector or a comma-separated string of subset of subject IDs}

\item{id_position}{the column position of the variable that unique identifies
the subjects}

\item{id_col_name}{the column name with the subject IDs}
}
\value{
a \code{list} of 2 elements of type \code{data.frame} that contain the project
data with only the records and fields of interest and its associated
metadata.
}
\description{
Subset records and columns from a REDCap project
}
\examples{
\dontrun{
  result <- redcap_read_rows_columns(
    base_url    = "https://bbmc.ouhsc.edu/redcap/api/",
    token       = "9A81268476645C4E5F03428B8AC3AA7B",
    fields      = c("record_id", "name_first", "age", "bmi"),
    records     = c("1", "3", "5"),
    id_position = 1L,
    id_col_name = NULL
  )
}

}
\keyword{internal}
