% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-request_parameters.R
\docType{data}
\name{request_parameters}
\alias{request_parameters}
\title{Request parameters}
\format{
A data frame of 3 rows and 12 columns
}
\usage{
request_parameters
}
\description{
The request_parameters data frame contains the information related to which
request parameters should be used when fetching data from a specific
version of  DHIS2. It currently contains parameters for versions ranging
from  2.22 to 2.42.
}
\keyword{datasets}
