% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{select_existing_tables}
\alias{select_existing_tables}
\title{Check if tables of interest are part of the tables in the database}
\usage{
select_existing_tables(table_names, tables, database_name)
}
\arguments{
\item{table_names}{a vector of table names of interest}

\item{tables}{the list of all tables in the database}

\item{database_name}{the database name}
}
\value{
the list of tables of interest that are part of the database
}
\description{
Check if tables of interest are part of the tables in the database
}
