% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_servers.R
\name{server_fetch_data_from_query}
\alias{server_fetch_data_from_query}
\title{Fetch data from server using an SQL query}
\usage{
server_fetch_data_from_query(src, tables, connexion)
}
\arguments{
\item{src}{the SQL query}

\item{tables}{the list of tables from the database}

\item{connexion}{the connexion object generated from the \code{authenticate()}
function}
}
\value{
a \code{list} of 1 or more objects of type \code{data.frame} containing each
the data fetched from a specific table.
}
\description{
Fetch data from server using an SQL query
}
\examples{
result <- server_fetch_data_from_query(
  src       = "select author_id, name, last_name from author",
  tables    = c("family_author", "author"),
  connexion = connexion
)
}
\keyword{internal}
