% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_servers.R
\name{server_fetch_data_from_table}
\alias{server_fetch_data_from_table}
\title{Fetch data from a server using the table name(s)}
\usage{
server_fetch_data_from_table(table_name, connexion, filter, select)
}
\arguments{
\item{table_name}{the name of the table in the server}

\item{connexion}{a connexion object created from the \code{authenticate()}
function}

\item{filter}{a vector or a comma-separated string of subset of subject IDs}

\item{select}{a vector of column names}
}
\value{
a \code{data.frame} with the data from the corresponding table name
}
\description{
Fetch data from a server using the table name(s)
}
\examples{
\dontrun{
  result <- server_fetch_data_from_table(
    table_names = "author",
    connexion   = connexion,
    filter      = filter,
    select      = select
  )
}
}
\keyword{internal}
