% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_servers.R
\name{server_find_table_and_query}
\alias{server_find_table_and_query}
\title{Identify queries or tables names from the user supplied string}
\usage{
server_find_table_and_query(src, tables)
}
\arguments{
\item{src}{the user supplied string. This is usually an SQL query or a table
name or a combination of both.}

\item{tables}{the list of all tables in the database}
}
\value{
a list with the identified queries and tables
}
\description{
Identify queries or tables names from the user supplied string
}
\keyword{internal}
