% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_servers.R
\name{server_make_query}
\alias{server_make_query}
\title{Make SQL query from the list of query parameters}
\usage{
server_make_query(table_name, connexion, filter, select)
}
\arguments{
\item{table_name}{The name of the table in the server}

\item{connexion}{A connection object created from the \code{authenticate()}
function}

\item{filter}{A vector or a comma-separated string of subset of subject IDs}

\item{select}{A vector of column names}
}
\value{
A string with the constructed SQL query from the provided query
parameter.
}
\description{
Make SQL query from the list of query parameters
}
\keyword{internal}
