% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_example_file.R
\name{show_example_file}
\alias{show_example_file}
\title{Display the structure of the credentials file}
\usage{
show_example_file()
}
\description{
Display the structure of the credentials file
}
\examples{
show_example_file()
}
