% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sormas-helpers.R
\name{sormas_get_api_specification}
\alias{sormas_get_api_specification}
\title{Download the API specification file from SORMAS}
\usage{
sormas_get_api_specification(path = tempdir(), overwrite = TRUE)
}
\arguments{
\item{path}{A character with the path to the folder where the downloaded data
dictionary should be stored. Default is \code{NULL} i.e. the data dictionary
will be stored in \code{tempdir()}}

\item{overwrite}{A logical used to specify whether to overwrite to overwrite
the existing data dictionary or not. Default is \code{TRUE}}
}
\value{
Invisibly returns the path to the folder where the file is stored.
When \code{path = NULL}, the file will be stored in the R temporary
folder as: \code{api_specification.json}
}
\description{
Download the API specification file from SORMAS
}
\examples{
# save the SORMAS API specification into the temporary R folder
path_api_specs <- sormas_get_api_specification()

}
