% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sormas-helpers.R
\name{sormas_get_data_dictionary}
\alias{sormas_get_data_dictionary}
\title{Download SORMAS data dictionary}
\usage{
sormas_get_data_dictionary(path = tempdir(), overwrite = TRUE)
}
\arguments{
\item{path}{A character with the path to the folder where the downloaded data
dictionary should be stored. Default is \code{NULL} i.e. the data dictionary
will be stored in \code{tempdir()}}

\item{overwrite}{A logical used to specify whether to overwrite to overwrite
the existing data dictionary or not. Default is \code{TRUE}}
}
\value{
A character with path to the folder where the data dictionary is
stored. When \code{path = NULL}, the file will be stored in the R
temporary folder as: \code{dictionary.xlsx}
}
\description{
Download SORMAS data dictionary
}
\examples{
# download and save the data dictionary in the default R temporary directory
path_to_dictionary <- sormas_get_data_dictionary()

}
