% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sormas-helpers.R
\name{sormas_get_diseases}
\alias{sormas_get_diseases}
\title{Get list of disease names from SORMAS}
\usage{
sormas_get_diseases(user_name, password)
}
\arguments{
\item{user_name}{The user name}

\item{password}{The user's password}
}
\value{
A vector of the list of disease names in SORMAS
}
\description{
Get list of disease names from SORMAS
}
\keyword{internal}
