% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sormas-helpers.R
\name{sormas_get_filters}
\alias{sormas_get_filters}
\title{Update data filtration options based on the user-provided parameters.}
\usage{
sormas_get_filters(filter)
}
\arguments{
\item{filter}{A list with the user-defined data filtration option}
}
\value{
A list with the data filtration options to be applied on the imported
data
}
\description{
Update data filtration options based on the user-provided parameters.
}
\keyword{internal}
