% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sormas-helpers.R
\name{sormas_get_pathogen_data}
\alias{sormas_get_pathogen_data}
\title{Get pathogen tests data from SORMAS}
\usage{
sormas_get_pathogen_data(base_url, user_name, password, since)
}
\arguments{
\item{base_url}{A character with the base URL to SORMAS system of interest}

\item{user_name}{A character with the user name}

\item{password}{A character with the user's password}

\item{since}{A Date value in ISO8601 format (YYYY-mm-dd). Default is \code{NULL}.}
}
\value{
A data frame with the following two columns: 'case_id', 'Ct_values'
}
\description{
Get pathogen tests data from SORMAS
}
\keyword{internal}
