% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sormas-helpers.R
\name{sormas_get_persons_data}
\alias{sormas_get_persons_data}
\title{Get personal data of cases from SORMAS}
\usage{
sormas_get_persons_data(base_url, user_name, password, since)
}
\arguments{
\item{base_url}{A character with the base URL to SORMAS system of interest}

\item{user_name}{A character with the user name}

\item{password}{A character with the user's password}

\item{since}{A Date value in ISO8601 format (YYYY-mm-dd). Default is \code{0} i.e.
to fetch all cases from the beginning of data collection.}
}
\value{
A data frame with the following eight columns: 'case_id',
'sex', 'date_of_birth', 'country', 'city', 'latitude', 'longitude'
}
\description{
Get personal data of cases from SORMAS
}
\keyword{internal}
