% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_connect_to_server}
\alias{sql_connect_to_server}
\title{Establish connection to the server}
\usage{
sql_connect_to_server(
  base_url,
  user_name,
  password,
  dbms,
  driver_name,
  database_name,
  port
)
}
\arguments{
\item{base_url}{the host server name}

\item{user_name}{the user name}

\item{password}{the user's password}

\item{dbms}{the database management system type}

\item{driver_name}{the driver name}

\item{database_name}{the database name}

\item{port}{the server port ID}
}
\value{
the \code{connection} object
}
\description{
Establish connection to the server
}
\examples{
\dontrun{
con <- sql_connect_to_server(
  base_url      = "mysql-rfam-public.ebi.ac.uk",
  user_name     = "rfamro",
  password      = "",
  dbms          = "MySQL",
  driver_name   = "",
  database_name = "Rfam",
  port          = 4497
)
}
}
\keyword{internal}
