% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_fetch_data_from_query}
\alias{sql_fetch_data_from_query}
\title{Fetch data from server using an SQL query}
\usage{
sql_fetch_data_from_query(
  src,
  tables,
  base_url,
  user_name,
  password,
  dbms,
  driver_name,
  database_name,
  port
)
}
\arguments{
\item{src}{the SQL query}

\item{tables}{the list of tables from the database}

\item{base_url}{the host server name}

\item{user_name}{the user name}

\item{password}{the user's password}

\item{dbms}{the database management system type}

\item{driver_name}{the driver name}

\item{database_name}{the database name}

\item{port}{the server port ID}
}
\value{
a \code{list} of 1 or more objects of type \code{data.frame} containing each
data fetched from one table.
}
\description{
Fetch data from server using an SQL query
}
\examples{
\dontrun{
result <- sql_fetch_data_from_query(
  src           = "select author_id, name, last_name from author",
  tables        = c("family_author", "author"),
  base_url      = "mysql-rfam-public.ebi.ac.uk",
  user_name     = "rfamro",
  password      = "",
  dbms          = "MySQL",
  driver_name   = "",
  database_name = "Rfam",
  port          = 4497
)
}
}
\keyword{internal}
