% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_select_entire_dataset}
\alias{sql_select_entire_dataset}
\title{Fetch entire dataset in a table}
\usage{
sql_select_entire_dataset(
  table,
  base_url,
  user_name,
  password,
  dbms,
  driver_name,
  database_name,
  port
)
}
\arguments{
\item{table}{the table name}

\item{base_url}{host server name}

\item{user_name}{the user name}

\item{password}{the user's password}

\item{dbms}{the database management system type}

\item{driver_name}{the driver name}

\item{database_name}{the database name}

\item{port}{the server port ID}
}
\value{
an object of type \code{data.frame} with the entire dataset fetched from
the specified table.
}
\description{
Fetch entire dataset in a table
}
\examples{
\dontrun{
result <- sql_select_entire_dataset(
  table         = "author",
  base_url      = "mysql-rfam-public.ebi.ac.uk",
  user_name     = "rfamro",
  password      = "",
  dbms          = "MySQL",
  driver_name   = "",
  database_name = "Rfam",
  port          = 4497
)
}
}
\keyword{internal}
