% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_select_fields_only}
\alias{sql_select_fields_only}
\title{Select specified fields from a table}
\usage{
sql_select_fields_only(
  table,
  base_url,
  user_name,
  password,
  dbms,
  driver_name,
  database_name,
  port,
  field
)
}
\arguments{
\item{table}{the table name}

\item{base_url}{the host server name}

\item{user_name}{the user name}

\item{password}{the user's password}

\item{dbms}{the database management system type}

\item{driver_name}{the driver name}

\item{database_name}{the database name}

\item{port}{the server port ID}

\item{field}{a vector of column names of interest}
}
\value{
an object of type \code{data.frame} that contains the data fetched from
the specific table with only the fields of interest.
}
\description{
Select specified fields from a table
}
\examples{
\dontrun{
result <- sql_select_fields_only(
  table         = "author",
  base_url      = "mysql-rfam-public.ebi.ac.uk",
  user_name     = "rfamro",
  password      = "",
  dbms          = "MySQL",
  driver_name   = "",
  database_name = "Rfam",
  port          = 4497,
  field         = c("author_id", "name", "last_name")
)
}
}
\keyword{internal}
