% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_select_records_only}
\alias{sql_select_records_only}
\title{Select specified records from a table}
\usage{
sql_select_records_only(
  table,
  base_url,
  user_name,
  password,
  dbms,
  driver_name,
  database_name,
  port = 4497L,
  record,
  id_column_name = NULL,
  id_pos = NULL
)
}
\arguments{
\item{table}{the table name}

\item{base_url}{the host server name}

\item{user_name}{the user name}

\item{password}{the user's password}

\item{dbms}{the database management system type}

\item{driver_name}{the driver name}

\item{database_name}{the database name}

\item{port}{the server port ID}

\item{record}{a vector or a comma-separated string of subset of subject IDs.}

\item{id_column_name}{the column names that unique identify the records in
the tables}

\item{id_pos}{a vector of the column positions of the variable that
unique identifies the subjects in each table}
}
\value{
an object of type \code{data.frame} that contains the data fetched from
the specific table with only the records of interest.
}
\description{
Select specified records from a table
}
\examples{
\dontrun{
result <- sql_select_records_only(
  table          = "author",
  base_url       = "mysql-rfam-public.ebi.ac.uk",
  user_name      = "rfamro",
  password       = "",
  dbms           = "MySQL",
  driver_name    = "",
  database_name  = "Rfam",
  port           = 4497,
  record         = c("1", "20", "50"),
  id_column_name = NULL,
  id_pos         = 1
)
}
}
\keyword{internal}
