% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers.R
\name{sql_server_read_data}
\alias{sql_server_read_data}
\title{Read data from database management systems (DBMS).}
\usage{
sql_server_read_data(
  user = "rfamro",
  password = "",
  host = "mysql-rfam-public.ebi.ac.uk",
  port = 1433L,
  database_name = "Rfam",
  driver_name = "",
  source = NULL,
  records = NULL,
  fields = NULL,
  id_position = NULL,
  id_col_name = NULL,
  dbms = "MySQL"
)
}
\arguments{
\item{user}{the user name}

\item{password}{the user password}

\item{host}{the name of the host server}

\item{port}{the port ID}

\item{database_name}{the name of the database that contains the table from
which the data should be pulled}

\item{driver_name}{the name of the MS driver. use \code{odbc::odbcListDrivers()}
to display the list of installed drivers}

\item{source}{an SQL query or a vector of table names from
the project or database. When this is not specified, the function will
extract data from all tables in the database.}

\item{records}{a vector or a comma-separated string of subset of subject IDs.
When specified, only the records that correspond to these subjects will
be imported.}

\item{fields}{a vector of strings where each string is a comma-separated list
of column names. The element of this vector should be a list of column
names from the first table specified in the \code{table_names} argument
and so on...}

\item{id_position}{a vector of the column positions of the variable that
unique identifies the subjects in each table. When the column name with
the subject IDs is known, use the \code{id_col_name} argument instead. default
is 1.}

\item{id_col_name}{the column name with the subject IDs}

\item{dbms}{the SQL server type}
}
\value{
a \code{list} of 1 or several objects of type \code{data.frame}. The number of
elements in the list depends on the number of tables from which the
data is fetched.
}
\description{
The function assumes the user has read access to the database.
Importing data stored in DBMS into R requires the installation
of the appropriate \code{driver} that is compatible with the server version
hosting the database. See the \code{vignette} for how to install the driver
}
\examples{
\dontrun{
data <- sql_server_read_data(
  user          = "rfamro",
  password      = "",
  host          = "mysql-rfam-public.ebi.ac.uk",
  port          = 4497,
  database_name = "Rfam",
  source        = "author",
  driver_name   = "",
  dbms          = "MySQL"
)
}
}
\keyword{internal}
