% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{subset_from_several_tables}
\alias{subset_from_several_tables}
\title{Subset data from several tables}
\usage{
subset_from_several_tables(result, records, id_col_name, table_names)
}
\arguments{
\item{result}{a list of data from the table name}

\item{records}{a vector or a comma-separated string of subset of subject IDs}

\item{id_col_name}{the column name with the subject IDs}

\item{table_names}{a vector of table names}
}
\value{
a list of data from the tables of interest with the records and
fields of interest
}
\description{
Subset data from several tables
}
