% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stuff.R
\name{subsetFields}
\alias{subsetFields}
\title{function to subset fields}
\usage{
subsetFields(data.frame, fields, table.name)
}
\arguments{
\item{data.frame}{the input data frame}

\item{fields}{the list of columns of interest}

\item{table.name}{the table names}
}
\value{
a list of 2 elements: the subset data frame and an integer that tells whether all fields were missing in the table (1) or not (0)
}
\description{
function to subset fields
}
\examples{
data = subsetFields(data.frame=data.frame, fields="id,data_entry_date", table.name="dss_events")
}
