% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stuff.R
\name{subsetRecords}
\alias{subsetRecords}
\title{function to subset records}
\usage{
subsetRecords(data.frame, records, id.position = 1, table.name)
}
\arguments{
\item{data.frame}{the input data frame}

\item{records}{the list of columns of interest}

\item{id.position}{a vector of the column position of the variable that unique identifies the subjects. If not provided, it will assumes the first column as the subject ID column in all the tables}

\item{table.name}{the table name}
}
\value{
a list of 2 elements: the subset data frame and an integer that tells whether all fields were missing in the table (1) or not (0)
}
\description{
function to subset records
}
\examples{
data = subsetRecords(data.frame, records, id.position=1, table.name)
}
