% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{url_check}
\alias{url_check}
\title{Check if the value for the base_url argument has a correct structure}
\usage{
url_check(base_url)
}
\arguments{
\item{base_url}{the URL to the HIS of interest}
}
\description{
Check if the value for the base_url argument has a correct structure
}
\keyword{internal}
