% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{url_check_domain_structure}
\alias{url_check_domain_structure}
\title{Check whether the domain structure of the URL is correct}
\usage{
url_check_domain_structure(url_parts)
}
\arguments{
\item{url_parts}{a character string with the domain from the URL}
}
\description{
Check whether the domain structure of the URL is correct
}
\keyword{internal}
