% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes.R
\name{check_dhis2_attributes}
\alias{check_dhis2_attributes}
\title{Check the DHIS2 attributes for data import}
\usage{
check_dhis2_attributes(
  username,
  password,
  base_url,
  dataset,
  organisation_unit = NULL,
  data_element_group = NULL
)
}
\arguments{
\item{username}{the user name}

\item{password}{the user's password}

\item{base_url}{the web address of the server the user wishes to log in to}

\item{dataset}{the dataSets identifiers}

\item{organisation_unit}{the organisationUnits identifiers}

\item{data_element_group}{the dataElementGroups identifiers}
}
\value{
a list of 7 elements of type \code{character}.
}
\description{
Check the DHIS2 attributes for data import
}
\examples{
\dontrun{
  attributes <- check_dhis2_attributes(
    username = "admin",
    password = "district",
    base_url = "https://play.dhis2.org/dev/",
    dataset = "pBOMPrpg1QX",
    organisation_unit = "DiszpKrYNg8",
    data_element_group = NULL
  )
}

}
