% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dhis2-helpers.R
\name{check_program}
\alias{check_program}
\title{Validate and retrieve program IDs}
\usage{
check_program(login, program)
}
\arguments{
\item{login}{A httr2 request object preconfigured for authentication
carrying the base url, user name, and password.}

\item{program}{A character of the program name to be validated}
}
\value{
The program ID if the provide program name or ID exists; otherwise,
an error is thrown with a suggestion to use the function that lists all
available programs from the target DHIS2 instance.
}
\description{
Checks whether the specified program ID or name is valid in a DHIS2 instance.
If the name provided is a valid program name or ID, it returns
the corresponding ID; otherwise, an error message is shown and a function
listing all available programs is provided.
}
\keyword{internal}
