% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_genomic_data.R
\name{clade_assignment}
\alias{clade_assignment}
\title{function to perform clade assignment using \strong{nextclade} and \strong{pangolin}}
\usage{
clade_assignment(merged.fasta, output.dir)
}
\arguments{
\item{merged.fasta}{path to the fasta file with the sequences from all samples}

\item{output.dir}{path to the folder where the clade assignment results will be stored}
}
\value{
the output files in the specified output directory
}
\description{
function to perform clade assignment using \strong{nextclade} and \strong{pangolin}
}
\details{
install both nextclade and pangolin using \strong{conda}
}
\examples{
clade_assignment(merged.fasta="/Users/karimmane/Documents/Karim/LSHTM/TRACE_dev/Genomic_data/Gam10_consensus/consensus_genomes.fasta", output.dir="/Users/karimmane/Documents/Karim/LSHTM/TRACE_dev/Genomic_data/Gam10_assignment")
}
