% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes-helpers.R
\name{dhis2_get_attributes}
\alias{dhis2_get_attributes}
\title{Get the target DHIS2 attribute identifiers and names}
\usage{
dhis2_get_attributes(base_url, username, password, which = "dataElements")
}
\arguments{
\item{base_url}{the base URL of the DHIS2 server}

\item{username}{the user name}

\item{password}{the user's password}

\item{which}{the target DHIS2 attribute name. Possible values are:
\enumerate{
\item dataSets: to get the dataset identifiers and names
\item organisationUnits: to get the organisation unit identifiers and names
\item dataElementGroups: to get the data element groups identifiers
and names
\item dataElements: to get the data elements identifiers and names
}}
}
\value{
an object of class \code{data.frame} that contains the information of
interest.
}
\description{
Get the target DHIS2 attribute identifiers and names
}
\keyword{internal}
