% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes_helpers.R
\name{dhis2_get_relevant_attributes}
\alias{dhis2_get_relevant_attributes}
\title{Get the relevant dataset}
\usage{
dhis2_get_relevant_attributes(
  attribute_id = NULL,
  base_url,
  username,
  password,
  which
)
}
\arguments{
\item{attribute_id}{a comma-separated list of DHIS2 attribute ids. The ids
could be those of a dataSet or an organisationUnit.}

\item{base_url}{the web address of the server the user wishes to log in to}

\item{username}{the user name}

\item{password}{the user's password}

\item{which}{the target DHIS2 attribute name}
}
\value{
a \code{list} of 2 elements: a \code{character} string with the target
attributes ID(s) and a \code{data.frame} that contains the data of interest
from the specified DHIS2 attribute ids.
}
\description{
Get the relevant dataset
}
\examples{
\dontrun{
result <- dhis2_get_relevant_attributes(
  attribute_id = "pBOMPrpg1QX,BfMAe6Itzgt",
  base_url = "https://play.dhis2.org/dev/",
  username = "admin",
  password = "district",
  which = "dataSets"
)
}
}
\keyword{internal}
