% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_dhis2-helpers.R
\name{dhis2_login}
\alias{dhis2_login}
\title{Check DHIS2 authentication details}
\usage{
dhis2_login(base_url, user_name, password)
}
\arguments{
\item{base_url}{the base URL of the DHIS2 server}

\item{user_name}{the user name}

\item{password}{the user's password}
}
\value{
a message if the dhis2_login was successful,
throws an error otherwise.

a message if the login was successful, throws an error otherwise.
}
\description{
If the user were granted with access to the API, this will return a message
specifying that the user was successfully connected. Otherwise, it will throw
an error message.
}
\examples{
\dontrun{
  dhis2_login(
    base_url  = file.path("https:/", "play.dhis2.org", "demo"),
    user_name = "admin",
    password  = "district"
  )
}
}
\keyword{internal}
