% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_dhis2-helpers.R
\name{dhis2_subset_fields}
\alias{dhis2_subset_fields}
\title{Subset fields when reading from DHIS2}
\usage{
<<<<<<< HEAD
dhis2_subset_fields(data, fields)
}
\arguments{
\item{data}{the input data frame}

\item{fields}{vector of fields to select from the data frame}
}
\value{
an object of type \code{data.frame} with the data that contains only the
fields of interest
=======
dhis2_subset_fields(fields, data)
}
\arguments{
\item{fields}{vector of fields to select from the data frame}

\item{data}{the input data frame}
}
\value{
the data frame with the fields of interest
>>>>>>> main
}
\description{
Subset fields when reading from DHIS2
}
\examples{
<<<<<<< HEAD
\dontrun{
results <- dhis2_subset_fields(
  data = readepi(
    credentials_file   = system.file("extdata", "test.ini",
    package = "readepi"),
    data_source        = "https://play.dhis2.org/dev",
    dataset            = "pBOMPrpg1QX,BfMAe6Itzgt",
    organisation_unit  = "DiszpKrYNg8",
    data_element_group = NULL,
    start_date         = "2014",
    end_date           = "2023",
    fields             = c("dataElement", "period", "value")
  )$data,
  fields               = c("dataElement", "period", "value")
)
}
}
\keyword{internal}
=======
results <- dhis2_subset_fields(
 fields = c("dataElement","period","value"),
 data = readepi(
  credentials_file = system.file("extdata", "test.ini", package = "readepi"),
   project_id = "DHIS2_DEMO",
   dataset = "pBOMPrpg1QX,BfMAe6Itzgt",
   organisation_unit = "DiszpKrYNg8",
   data_element_group = NULL,
   start_date = "2014",
   end_date = "2023",
   fields = c("dataElement","period","value")
 )$data
)
}
>>>>>>> main
