% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{dhis2_subset_fields}
\alias{dhis2_subset_fields}
\title{Subset fields when reading from DHIS2}
\usage{
dhis2_subset_fields(fields, data)
}
\arguments{
\item{fields}{vector of fields to select from the data frame}

\item{data}{the input data frame}
}
\value{
the data frame with the fields of interest
}
\description{
Subset fields when reading from DHIS2
}
