% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{dhis2_subset_records}
\alias{dhis2_subset_records}
\title{Title}
\usage{
dhis2_subset_records(records, id_col_name, data)
}
\arguments{
\item{records}{a vector of records to select from the data}

\item{id_col_name}{the column name where the records belong to}

\item{data}{the input data frame}
}
\value{
a data frame with the records of interest
}
\description{
Title
}
