% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_dhis2_helpers.R
\name{dhis2_subset_records}
\alias{dhis2_subset_records}
\title{Subset a specified set of records from a dataset imported from DHIS2}
\usage{
dhis2_subset_records(data, records, id_col_name)
}
\arguments{
\item{data}{the input data frame}

\item{records}{a vector of records to select from the data}

\item{id_col_name}{the column name where the records belong to}
}
\value{
an object of type \code{data.frame} with the data that contains only the
records of interest
}
\description{
Subset a specified set of records from a dataset imported from DHIS2
}
\examples{
\dontrun{
result <- dhis2_subset_records(
  data = readepi(
    credentials_file = system.file("extdata", "test.ini",
      package = "readepi"
    ),
    project_id = "DHIS2_DEMO",
    dataset = "pBOMPrpg1QX",
    organisation_unit = "DiszpKrYNg8",
    data_element_group = NULL,
    start_date = "2014",
    end_date = "2023"
  )$data,
  records = c("FTRrcoaog83", "eY5ehpbEsB7", "Ix2HsbDMLea"),
  id_col_name = "dataElement"
)
}
}
\keyword{internal}
