% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{fetch_data_from_query}
\alias{fetch_data_from_query}
\title{Fetch data from server using an SQL query}
\usage{
fetch_data_from_query(
  source,
  dbms,
  tables,
  driver_name,
  host,
  database_name,
  user,
  password,
  port
)
}
\arguments{
\item{source}{the SQL query}

\item{dbms}{the database management system type}

\item{tables}{the list of tables from the database}

\item{driver_name}{the driver name}

\item{host}{the host server name}

\item{database_name}{the database name}

\item{user}{the user name}

\item{password}{the user's password}

\item{port}{the server port ID}
}
\value{
a \code{list} of 1 or more objects of type \code{data.frame} containing each
data fetched from one table.
}
\description{
Fetch data from server using an SQL query
}
\examples{
\dontrun{
  result <- fetch_data_from_query(
    source = "select author_id, name, last_name from author",
    dbms = "MySQL",
    tables = c("family_author", "author"),
    driver_name = "",
    host = "mysql-rfam-public.ebi.ac.uk",
    database_name = "Rfam",
    user = "rfamro",
    password = "",
    port = 4497
  )
}

}
