% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips-helpers.R
\name{fingertips_subset_columns}
\alias{fingertips_subset_columns}
\title{Subset columns when reading from Fingertips}
\usage{
fingertips_subset_columns(fields, data)
}
\arguments{
\item{fields}{a vector or a comma-separated string of column names}

\item{data}{the data read from Fingertips}
}
\value{
a data frame with the columns of interest
}
\description{
Subset columns when reading from Fingertips
}
\examples{
\dontrun{
res <- fingertips_subset_columns(
  fields = c("IndicatorID", "AreaCode", "Age", "Value"),
  data = readepi(
    profile_id = 19,
    area_type_id = 202
  )$data
)
}

}
