% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips_helpers.R
\name{fingertips_subset_columns}
\alias{fingertips_subset_columns}
\title{Subset columns when reading from Fingertips}
\usage{
fingertips_subset_columns(data, fields)
}
\arguments{
\item{data}{the data read from Fingertips}

\item{fields}{a vector or a comma-separated string of column names}
}
\value{
an object of type \code{data.frame} with the Fingertips dataset that
contains only the fields of interest
}
\description{
Subset columns when reading from Fingertips
}
\examples{
\dontrun{
res <- fingertips_subset_columns(
  data = readepi(
    profile_id   = 19,
    area_type_id = 202
  )$data,
  fields         = c("IndicatorID", "AreaCode", "Age", "Value")
)
}
}
\keyword{internal}
