% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips_helpers.R
\name{fingertips_subset_rows}
\alias{fingertips_subset_rows}
\title{Subset records when reading from Fingertips}
\usage{
fingertips_subset_rows(data, records, id_col_name)
}
\arguments{
\item{data}{the data read from Fingertips}

\item{records}{a vector or a comma-separated string of records}

\item{id_col_name}{the column name with the subject IDs}
}
\value{
an object of type \code{data.frame} with the Fingertips dataset that
contains only the records of interest
}
\description{
Subset records when reading from Fingertips
}
\examples{
\dontrun{
res <- fingertips_subset_rows(
  data = readepi(
    profile_id = 19,
    area_type_id = 202
  )$data,
  records = c("E92000001", "E12000002", "E12000009"),
  id_col_name = "AreaCode"
)
}

}
