% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_data_element_groups}
\alias{get_data_element_groups}
\title{Get the data element group identifiers and names}
\usage{
get_data_element_groups(base.url, username, password)
}
\arguments{
\item{base.url}{the base URL of the DHIS2 server}

\item{username}{the user name}

\item{password}{the user's password}
}
\description{
Get the data element group identifiers and names
}
