% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_data_element_groups}
\alias{get_data_element_groups}
\title{function to get the data element group identifiers and names}
\usage{
get_data_element_groups(base.url, username, password)
}
\description{
function to get the data element group identifiers and names
}
