% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_data_elements}
\alias{get_data_elements}
\title{function to get the data element identifiers and names}
\usage{
get_data_elements(base.url, username, password)
}
\description{
function to get the data element identifiers and names
}
