% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes-helpers.R
\name{get_data_elements}
\alias{get_data_elements}
\title{Get the data element identifiers and names}
\usage{
get_data_elements(base_url, username, password)
}
\arguments{
\item{base_url}{the base URL of the DHIS2 server}

\item{username}{the user name}

\item{password}{the user's password}
}
\description{
Get the data element identifiers and names
}
