% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_data_sets}
\alias{get_data_sets}
\title{function to get the datasets identifiers and names}
\usage{
get_data_sets(base.url, username, password)
}
\description{
function to get the datasets identifiers and names
}
