% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_file-helpers.R
\name{get_extension}
\alias{get_extension}
\title{Get file extension}
\usage{
get_extension(file_path)
}
\arguments{
\item{file_path}{the target file path}
}
\value{
a \code{character} that contains the file extension
}
\description{
Get file extension
}
\examples{
\dontrun{
  ext <- get_extension(
    file_path = system.file("extdata", "test.txt", package = "readepi")
  )
}

}
