% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips-helpers.R
\name{get_fingertips_metadata}
\alias{get_fingertips_metadata}
\title{get fingertips metadata}
\usage{
get_fingertips_metadata()
}
\value{
a list of data frames
}
\description{
get fingertips metadata
}
\examples{
metadata <- get_fingertips_metadata()

}
