% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips_helpers.R
\name{get_fingertips_metadata}
\alias{get_fingertips_metadata}
\title{get fingertips metadata}
\usage{
get_fingertips_metadata()
}
\value{
a \code{list} of 3 objects of type \code{data.frame}. They contain information
about the indicators in the Fingertips repository.
}
\description{
get fingertips metadata
}
\examples{
\dontrun{
metadata <- get_fingertips_metadata()
}
}
\keyword{internal}
