% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{get_id_column_name}
\alias{get_id_column_name}
\title{get the id column name}
\usage{
get_id_column_name(id_col_name, j, id_position)
}
\arguments{
\item{id_col_name}{the id column name}

\item{j}{the index}

\item{id_position}{the id position}
}
\description{
get the id column name
}
