% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips-helpers.R
\name{get_ind_id_from_domain_id}
\alias{get_ind_id_from_domain_id}
\title{Get indicator ID from domain ID}
\usage{
get_ind_id_from_domain_id(metadata, domain_id, indicator_name = NULL)
}
\arguments{
\item{metadata}{a list with the fingertips metadata}

\item{domain_id}{the domain ID}

\item{indicator_name}{the indicator name}
}
\value{
the indicator ID
}
\description{
Get indicator ID from domain ID
}
\examples{
indicator_id <- get_ind_id_from_domain_id(
metadata = list(
  indicator_profile_domain = fingertipsR::indicators(),
  indicator_ids_names = fingertipsR::indicators_unique(),
  area_type = fingertipsR::area_types()
  ),
domain_id = 1000041,
indicator_name = NULL
)

}
