% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips-helpers.R
\name{get_ind_id_from_domain_name}
\alias{get_ind_id_from_domain_name}
\title{Get indicator ID from domain name}
\usage{
get_ind_id_from_domain_name(metadata, domain_name, indicator_name = NULL)
}
\arguments{
\item{metadata}{a list with the fingertips metadata}

\item{domain_name}{the domain name}

\item{indicator_name}{the indicator name}
}
\value{
the indicator ID
}
\description{
Get indicator ID from domain name
}
