% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips-helpers.R
\name{get_ind_id_from_ind_name}
\alias{get_ind_id_from_ind_name}
\title{Get indicator ID from indicator name}
\usage{
get_ind_id_from_ind_name(metadata, indicator_name)
}
\arguments{
\item{metadata}{a list with the fingertips metadata}

\item{indicator_name}{the indicator name}
}
\value{
the indicator ID
}
\description{
Get indicator ID from indicator name
}
\examples{
\dontrun{
indicator_id <- get_ind_id_from_ind_name(
metadata = list(
  indicator_profile_domain = fingertipsR::indicators(),
  indicator_ids_names = fingertipsR::indicators_unique(),
  area_type = fingertipsR::area_types()
  ),
indicator_name = "Pupil absence"
)
}


}
