% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_organisation_units}
\alias{get_organisation_units}
\title{function to get the organisation unit identifiers and names}
\usage{
get_organisation_units(base.url, username, password)
}
\description{
function to get the organisation unit identifiers and names
}
