% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_fingertips-helpers.R
\name{get_profile_name}
\alias{get_profile_name}
\title{Get profile name from Fingertips}
\usage{
get_profile_name(profile_id, profile_name, metadata)
}
\arguments{
\item{profile_id}{the profile ID}

\item{profile_name}{the profile name}

\item{metadata}{the Fingertips metadata}
}
\value{
a list with the profile name and their correspondent indexes
}
\description{
Get profile name from Fingertips
}
\examples{
res <- get_profile_name(
profile_id = 19,
profile_name = "Public Health Outcomes Framework",
metadata = list(
  indicator_profile_domain = fingertipsR::indicators(),
  indicator_ids_names = fingertipsR::indicators_unique(),
  area_type = fingertipsR::area_types()
  )
)

}
