% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_read_file_args}
\alias{get_read_file_args}
\title{Get arguments for reading from files}
\usage{
get_read_file_args(args_list)
}
\arguments{
\item{args_list}{a \code{list} of user specified arguments}
}
\value{
a \code{list} of 4 elements of types \code{characters}. They define how the
file will be read.
}
\description{
Get arguments for reading from files
}
\examples{
\dontrun{
args_list <- get_read_file_args(
 list(
   sep = "\t",
   format = ".txt",
   which = NULL
 )
)
}

}
