% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{get_read_fingertips_args}
\alias{get_read_fingertips_args}
\title{Get arguments for reading from Fingertips}
\usage{
get_read_fingertips_args(args_list)
}
\arguments{
\item{args_list}{a list of user specified arguments}
}
\value{
a list of the parameters to be used for reading from Fingertips
}
\description{
Get arguments for reading from Fingertips
}
\examples{
\dontrun{
args_list <- get_read_fingertips_args(
 list(
   indicator_id = 90362,
   area_type_id = 202
 )
)
}

}
