% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes-helpers.R
\name{get_relevant_dataset}
\alias{get_relevant_dataset}
\title{Get the relevant dataset}
\usage{
get_relevant_dataset(dataset, base_url, username, password)
}
\arguments{
\item{dataset}{the dataSets identifiers}

\item{base_url}{the web address of the server the user wishes to log in to}

\item{username}{the user name}

\item{password}{the user's password}
}
\value{
a \code{list} of 2 elements: a \code{character} string with the target dataset
ID(s) and a \code{data.frame} of the details about all datasets in the target
DHIS2 server.
}
\description{
Get the relevant dataset
}
\examples{
\dontrun{
  result <- get_relevant_dataset(
    dataset = "pBOMPrpg1QX,BfMAe6Itzgt",
    base_url = "https://play.dhis2.org/dev/",
    username = "admin",
    password = "district"
  )
}

}
