% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dhis2_attributes-helpers.R
\name{get_relevant_organisation_unit}
\alias{get_relevant_organisation_unit}
\title{Get the relevant organisation units}
\usage{
get_relevant_organisation_unit(organisation_unit, base_url, username, password)
}
\arguments{
\item{organisation_unit}{the organisationUnits identifiers}

\item{base_url}{the web address of the server the user wishes to log in to}

\item{username}{the user name}

\item{password}{the user's password}
}
\value{
a \code{list} of 2 elements: a \code{character} string with the target
organisation unit ID(s) and a \code{data.frame} of the details about all
organisation unit in the target DHIS2 server.
}
\description{
Get the relevant organisation units
}
\examples{
\dontrun{
  result <- get_relevant_organisation_unit(
    organisation_unit = "DiszpKrYNg8",
    base_url = "https://play.dhis2.org/dev/",
    username = "admin",
    password = "district"
  )
}

}
