% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{identify_table_name}
\alias{identify_table_name}
\title{Detect table names from an SQL query}
\usage{
identify_table_name(query, tables)
}
\arguments{
\item{query}{the SQL query}

\item{tables}{the list of all tables from the database}
}
\value{
the table name of interest
}
\description{
Detect table names from an SQL query
}
\examples{
table_name = identify_table_name(
 query = "select * from author",
 tables = c("family_author", "author", "test")
)

}
