% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_redcap-helpers.R
\name{import_redcap_data}
\alias{import_redcap_data}
\title{Import data from REDCap under all scenari}
\usage{
import_redcap_data(records, fields, uri, token, id_position, id_col_name)
}
\arguments{
\item{records}{a vector or a comma-separated string of subset of subject IDs}

\item{fields}{a vector or a comma-separated string of column names}

\item{uri}{the URI of the server}

\item{token}{the user-specific string that serves as the password for a
project}

\item{id_position}{the column position of the variable that unique identifies
the subjects}

\item{id_col_name}{the column name with the subject IDs}
}
\value{
a list with the data of interest and its associated metadata
}
\description{
This is a wrapper across all the use case of reading data from
REDCap i.e. around the function that all records and fields from the project,
around the function that read specific records/fields or both at the same
time
}
\examples{
result = import_redcap_data(
 uri = "https://bbmc.ouhsc.edu/redcap/api/",
 token = "9A81268476645C4E5F03428B8AC3AA7B",
 records = c("1", "3", "5"),
 fields = c("record_id", "name_first", "age", "bmi"),
 id_col_name = NULL,
 id_position = 1
)
}
