% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dep.R
\name{install_odbc_driver}
\alias{install_odbc_driver}
\title{Install the Microsoft ODBC driver for SQL server}
\usage{
install_odbc_driver(driver_version = 17)
}
\arguments{
\item{driver_version}{the version of the driver that is compatible with your SQL server}
}
\description{
Install the Microsoft ODBC driver for SQL server
}
\examples{
\dontrun{
install_odbc_driver(driver.version = 17)
}
}
