% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{login}
\alias{login}
\title{Check DHIS2 authentication details}
\usage{
login(username, password, base.url)
}
\arguments{
\item{username}{the user name}

\item{password}{the user's password}

\item{base.url}{the base URL of the DHIS2 server}
}
\description{
Check DHIS2 authentication details
}
