% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_dhis2.R
\name{read_from_dhis2}
\alias{read_from_dhis2}
\title{Import data form DHIS2 into R}
\usage{
read_from_dhis2(
  base_url,
  user_name,
  password,
  dataset,
  organisation_unit,
  data_element_group,
  start_date,
  end_date,
  records = NULL,
  fields = NULL,
  id_col_name = "dataElement"
)
}
\arguments{
\item{base_url}{the web address of the server the user wishes to log in to}

\item{user_name}{the user name}

\item{password}{the user password}

\item{dataset}{a vector or a list of comma-separated data set identifiers}

\item{organisation_unit}{a vector or a list of comma-separated organisation
unit identifiers}

\item{data_element_group}{a vector or a list of comma-separated data element
group identifiers}

\item{start_date}{the start date for the time span of the values to export}

\item{end_date}{the end date for the time span of the values to export}

\item{records}{a vector or a comma-separated string of subset of subject IDs.
When specified, only the records that correspond to these subjects will be
imported.}

\item{fields}{a vector or a comma-separated string of column names.
If provided, only those columns will be imported.}

\item{id_col_name}{the column name with the records of interest.}
}
\value{
a \code{list} of 1 element of type \verb{data frame}.
}
\description{
Import data form DHIS2 into R
}
\examples{
\dontrun{
data <- read_from_dhis2(
 base_url = file.path("https:/", "play.dhis2.org", "dev"),
 user_name = "admin",
 password = "district",
 dataset = "pBOMPrpg1QX",
 organisation_unit = "DiszpKrYNg8",
 data_element_group = "oDkJh5Ddh7d",
 start_date = "2014",
 end_date = "2023",
 records = NULL,
 fields = NULL,
 id_col_name = "dataElement"
)
}
}
\keyword{internal}
