% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_dhis2.R
\name{read_from_dhis2}
\alias{read_from_dhis2}
\title{Import data form DHIS2 into R}
\usage{
read_from_dhis2(base.url, user, password, records = NULL, fields = NULL)
}
\arguments{
\item{base.url}{the web address of the server the user wishes to log in to}

\item{user}{the user name}

\item{password}{the user password}

\item{records}{a vector or a comma-separated string of subset of subject IDs. When specified, only the records that correspond to these subjects will be imported.}

\item{fields}{a vector or a comma-separated string of column names. If provided, only those columns will be imported.}
}
\value{
a data frame
}
\description{
Import data form DHIS2 into R
}
