% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stuff.R
\name{read_from_ms_sql_server}
\alias{read_from_ms_sql_server}
\title{function to read data from relational databases hosted by online MS SQL server. The user needs to have read access to the database et should install the appropriate MS driver that is compatible with the SQLServer version}
\usage{
read_from_ms_sql_server(
  user,
  password,
  host,
  port,
  database.name,
  driver.name,
  table.name,
  records = NULL,
  fields = NULL,
  id.position = 1
)
}
\arguments{
\item{user}{the user name}

\item{password}{the user password}

\item{host}{the name of the host server}

\item{port}{the port ID}

\item{database.name}{the name of the database that contains the table from which the data should be pulled}

\item{driver.name}{the name of the MS driver. use \code{odbc::odbcListDrivers()} to display the list of installed drivers}

\item{table.name}{the name of the target table}

\item{records}{a vector or a comma-separated string of subset of subject IDs. When specified, only the records that correspond to these subjects will be imported.}

\item{fields}{a vector or a comma-separated string of column names. If provided, only those columns will be imported}

\item{id.position}{the column position of the variable that unique identifies the subjects. This should only be specified when the column with the subject IDs is not the first column. default is 1.}
}
\value{
a data frame
}
\description{
function to read data from relational databases hosted by online MS SQL server. The user needs to have read access to the database et should install the appropriate MS driver that is compatible with the SQLServer version
}
\examples{
data = read_from_relationalDB(user="kmane", password="****", host="robin.mrc.gm", port=1433, database.name="mrcg_db", table.name="test", driver.name="ODBC Driver 17 for SQL Server")
}
