% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_redcap.R
\name{read_from_redcap}
\alias{read_from_redcap}
\title{Import data from REDCap into R}
\usage{
read_from_redcap(
  uri,
  token,
  id.position = 1,
  id.col.name = NULL,
  records = NULL,
  fields = NULL
)
}
\arguments{
\item{uri}{the URI of the server}

\item{token}{the user-specific string that serves as the password for a project}

\item{id.position}{the column position of the variable that unique identifies the subjects. When the column name with the subject IDs is known, use the \code{id.col.name} argument instead. default is 1}

\item{id.col.name}{the column name with the subject IDs}

\item{records}{a vector or a comma-separated string of subset of subject IDs. When specified, only the records that correspond to these subjects will be imported.}

\item{fields}{a vector or a comma-separated string of column names. If provided, only those columns will be imported.}
}
\value{
a list with 2 data frames: the data of interest and the metadata associated to the data.
}
\description{
Import data from REDCap into R
}
\examples{
\dontrun{
redcap.data <- read_from_redcap(
  uri = "https://redcap.mrc.gm:8443/redcap/api/",
  token = "****",
  id.position = 1,
  id.col.name = NULL,
  records = NULL,
  fields = NULL
)
}
}
