% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{read_multiple_files}
\alias{read_multiple_files}
\title{Read multiple files, including multiple files in a directory}
\usage{
read_multiple_files(files, dirs, format = NULL, which = NULL)
}
\arguments{
\item{files}{a file or vector of file path to import}

\item{dirs}{a directory or a vector of directories where files will be
imported from}

\item{format}{a string used to specify the file format. This is useful when a
file does not have an extension, or has a file extension that does not match
its actual type}

\item{which}{a string used to specify the name of the excel sheet to import}
}
\value{
a list of data frames where each data frame contains data from a file
}
\description{
Read multiple files, including multiple files in a directory
}
