% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_file_helpers.R
\name{read_rio_formats}
\alias{read_rio_formats}
\title{Read files using the \code{rio} package}
\usage{
read_rio_formats(files_extensions, rio_extensions, files, files_base_names)
}
\arguments{
\item{files_extensions}{a vector a file extensions made from your files of
interest}

\item{rio_extensions}{a vector of files extensions supported by the rio
package}

\item{files}{a vector a files of interest}

\item{files_base_names}{a vector of file base}
}
\value{
a \code{list} of 4 elements of type \code{character} and \code{list}.
}
\description{
Read files using the \code{rio} package
}
