% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_servers.R
\name{read_server}
\alias{read_server}
\title{Read data from database management systems (DBMS).}
\usage{
read_server(conn, query)
}
\arguments{
\item{conn}{The connection object obtained from the \code{authenticate()}
function.}

\item{query}{An SQL query or a list with the following elements:
\enumerate{
\item table: a string with the table name
\item select: a vector of column names. When specified, only
those columns will be returned. Default is \code{NULL}.
\item filter: an expression or a vector of values used to filter the
rows from the table of interest. This should be of the same length
as the value for the 'select'. Default is \code{NULL}.
}}
}
\value{
A \code{data.frame} with the requested data as specified in the \code{query}
argument.
}
\description{
The function assumes the user has read access to the database.
Importing data from DBMS requires the installation of the appropriate
\code{driver} that is compatible with the server version hosting the database.
For more details, see the \href{../vignettes/install_drivers.Rmd}{vignette} on
how to install the driver.
}
\examples{
\dontrun{
  # establish the connection to the database
  conn <- authenticate(
    from          = "mysql-rfam-public.ebi.ac.uk",
    type          = "MySQL",
    user_name     = "rfamro",
    password      = "",
    driver_name   = "",
    db_name       = "Rfam",
    port          = 4497
  )

  # import data where query parameters are specified as a list
  authors_list <- read_server(
    conn  = conn,
    query = list(table = "author", select = NULL, filter = NULL)
  )

  # import data where query parameters is within an SQL query
  authors_list <- read_server(
    conn  = conn,
    query = "select * from author"
  )
}
}
