% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sormas.R
\name{read_sormas}
\alias{read_sormas}
\title{Import data from SORMAS}
\usage{
read_sormas(base_url, user_name, password, disease, since = NULL)
}
\arguments{
\item{base_url}{A character with the base URL to SORMAS system of interest}

\item{user_name}{A character with the user name}

\item{password}{A character with the user's password}

\item{disease}{A character with the target disease name}

\item{since}{A Date value in ISO8601 format (YYYY-mm-dd). Default is \code{NULL}.}
}
\value{
A data frame with the case data of the specified disease.
}
\description{
The function returns the following columns by default: \code{case_id,
person_id, sex, date_of_birth, case_origin, country, city, lat, long,
case_status, date_onset, date_admission, date_last_contact,
date_first_contact, outcome, date_outcome, Ct_values}.
}
\examples{
# fetch all COVID (coronavirus) cases from the test SORMAS instance
covid_cases <- read_sormas(
  base_url = "https://demo.sormas.org/sormas-rest",
  user_name = "SurvSup",
  password = "Lk5R7JXeZSEc",
  disease = "coronavirus",
)

}
