% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sormas.R
\name{read_sormas}
\alias{read_sormas}
\title{Import data from SORMAS (Surveillance, Outbreak Response Management and
Analysis System)}
\usage{
read_sormas(user_name, password, query_parameters)
}
\arguments{
\item{user_name}{The user name}

\item{password}{The user's password}

\item{query_parameters}{A list with the query parameters. Possible values
are:
\enumerate{
\item disease: a character with the target disease name (required).
\item uuid: a character with the target case id. Default is "all" to
fetch all cases of the specified disease (required).
\item since: a numeric used to specify the target date. Only cases
reported from this date till now will be returned (required).
Default is 0 to fetch cases from the first date.
When a Date value is provided instead of a numeric value, this
will be converted internally.
}}
}
\value{
A data frame with the case data of the specified disease.
}
\description{
Import data from SORMAS (Surveillance, Outbreak Response Management and
Analysis System)
}
\examples{
# fetch all COVID (coronavirus) cases from the first date
covid_cases <- read_sormas(
  user_name = "SurvSup",
  password = "Lk5R7JXeZSEc",
  query_parameters <- list(
    disease = "coronavirus",
    uuid = "all", # could take a vector of uuids
    since = 0 # cases from this date to now will be fetched
  )
)

}
