% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stuff.R
\name{readCredentials}
\alias{readCredentials}
\title{function to read credentials from a configuration file}
\usage{
readCredentials(file.path, project.id)
}
\description{
function to read credentials from a configuration file
}
