% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi.R
\name{readepi}
\alias{readepi}
\title{Import data from different sources into R}
\usage{
readepi(
  credentials.file = NULL,
  file.path = NULL,
  sep = NULL,
  format = NULL,
  which = NULL,
  pattern = NULL,
  project.id = NULL,
  driver.name = NULL,
  table.name = NULL,
  records = NULL,
  fields = NULL,
  id.position = NULL,
  dataset = NULL,
  organisation.unit = NULL,
  data.element.group = NULL,
  start.date = NULL,
  end.date = NULL,
  id.col.name = NULL
)
}
\arguments{
\item{credentials.file}{the path to the file with the user-specific credential details
for the projects of interest. It is required when importing data from DBMS. This is a
tab-delimited file with the following columns:
\enumerate{
\item user_name: the user name
\item password: the user password (for REDCap, this corresponds to the \strong{token} that serves as password to the project)
\item host_name: the host name (for MS SQL servers) or the URI (for REDCap)
\item project_id: the project ID or the name of the database you are access to.
\item comment: a summary description about the project or database of interest
\item dbms: the name of the DBMS: 'redcap' or 'REDCap' when reading from REDCap, 'sqlserver' or 'SQLServer' when reading from MS SQLServer
\item port: the port ID (used for MS SQLServers)
}
Use the \code{show_example_file()} function to display the structure of the template
credentials file}

\item{file.path}{the path to the file to be read. When several files need to be imported from a directory, this should be the path to that directory}

\item{sep}{the separator between the columns in the file. This is only required for space-separated files.}

\item{format}{a string used to specify the file format. This is useful when a file does not have an extension, or has a file extension that does not match its actual type}

\item{which}{a string used to specify which objects should be extracted (the name of the excel sheet to import)}

\item{pattern}{when specified, only files with this suffix will be imported from the specified directory}

\item{project.id}{for relational DB, this is the name of the database that contains the table from which the data should be pulled. Otherwise, it is the project ID you were given access to.}

\item{driver.name}{the name of the MS driver. use \code{odbc::odbcListDrivers()} to display the list of installed drivers}

\item{table.name}{the name of the target table}

\item{records}{a vector or a comma-separated string of subset of subject IDs. When specified, only the records that correspond to these subjects will be imported.}

\item{fields}{a vector or a comma-separated string of column names. If provided, only those columns will be imported.}

\item{id.position}{the column position of the variable that unique identifies the subjects. When the name of the column with the subject IDs is known, this can be provided using the \code{id.col.name} argument}

\item{dataset}{a vector or a list of comma-separated data set identifiers.}

\item{organisation.unit}{a vector or a list of comma-separated organisation unit identifiers.}

\item{data.element.group}{a vector or a list of comma-separated data element group identifiers.}

\item{start.date}{the start date for the time span of the values to export}

\item{end.date}{the end date for the time span of the values to export}

\item{id.col.name}{the column name with the subject IDs.}
}
\value{
a list with 2 data frames (data and metadata) when reading from REDCap. A data frame otherwise.
}
\description{
the function allows import of data from files or health information systems (HIS)
The HIS are database management systems (DBMS), both relational and NoSQL databases.
}
\examples{
\dontrun{
# reading from a MS SQL server
data <- readepi(
  credentials.file = system.file("extdata", "test.ini", package = "readepi"),
  project.id = "IBS_BHDSS",
  driver.name = "ODBC Driver 17 for SQL Server",
  table.name = "dss_events"
)
}
}
