% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stuff.R
\name{readMSsqlCredentials}
\alias{readMSsqlCredentials}
\title{function to read user credentials from file}
\usage{
readMSsqlCredentials(credentials.file)
}
\description{
function to read user credentials from file
}
