% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_redcap_helpers.R
\name{redcap_get_results}
\alias{redcap_get_results}
\title{Check and return data fetch from redcap}
\usage{
redcap_get_results(redcap_data, metadata)
}
\arguments{
\item{redcap_data}{the object with redcap data}

\item{metadata}{the object with redcap metadata}
}
\value{
a \code{list} of 2 elements of type \code{data.frame}. These are the dataset of
interest and its associated metadata.
}
\description{
Check and return data fetch from redcap
}
\examples{
\dontrun{
result <- redcap_get_results(
  redcap_data = REDCapR::redcap_read(
    redcap_uri = "https://bbmc.ouhsc.edu/redcap/api/",
    token = "9A81268476645C4E5F03428B8AC3AA7B",
    records = c("1", "3", "5"),
    fields = c("record_id", "name_first", "age", "bmi"),
    verbose = FALSE,
    id_position = 1L
  ),
  metadata = REDCapR::redcap_metadata_read(
    redcap_uri = "https://bbmc.ouhsc.edu/redcap/api/",
    token = "9A81268476645C4E5F03428B8AC3AA7B",
    fields = NULL,
    verbose = FALSE
  )
)
}

}
