% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{redcap_read_records}
\alias{redcap_read_records}
\title{Subset records from a REDCap project}
\usage{
redcap_read_records(records, uri, token, id_position, id_col_name)
}
\arguments{
\item{records}{a vector or a comma-separated string of subset of subject IDs}

\item{uri}{the URI of the REDCap project}

\item{token}{the user-specific string that serves as the password for a
project}

\item{id_position}{the column position of the variable that unique identifies
the subjects}

\item{id_col_name}{the column name with the subject IDs}
}
\value{
a list with the project data and its associated metadata with the
records of interest
}
\description{
Subset records from a REDCap project
}
