% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_redcep_helpers.R
\name{redcap_read_rows_columns}
\alias{redcap_read_rows_columns}
\title{Subset records and columns from a REDCap project}
\usage{
redcap_read_rows_columns(fields, uri, token, id_position, id_col_name, records)
}
\arguments{
\item{fields}{a vector or a comma-separated string of column names}

\item{uri}{the URI of the REDCap project}

\item{token}{the user-specific string that serves as the password for a
project}

\item{id_position}{the column position of the variable that unique identifies
the subjects}

\item{id_col_name}{the column name with the subject IDs}

\item{records}{a vector or a comma-separated string of subset of subject IDs}
}
\value{
a \code{list} of 2 elements of type \code{data.frame} that contain the project
data with only the records and fields of interest and
its associated metadata.
}
\description{
Subset records and columns from a REDCap project
}
