% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rdbms-helpers.R
\name{server_fetch_data}
\alias{server_fetch_data}
\title{Fetch all or specific rows and columns from a table}
\usage{
server_fetch_data(login, query)
}
\arguments{
\item{login}{The connection object obtained from the \code{login()} function.}

\item{query}{A string with the SQL query}
}
\value{
A data frame with the data of interest
}
\description{
Fetch all or specific rows and columns from a table
}
\keyword{internal}
