% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rdbms-helpers.R
\name{server_make_query}
\alias{server_make_query}
\title{Make an SQL query from a list of query parameters}
\usage{
server_make_query(table_name, login, filter, select)
}
\arguments{
\item{table_name}{The name of the table in the server
function}

\item{login}{The connection object obtained from the \code{login()} function.}

\item{filter}{A vector or a comma-separated string of subset of subject IDs}

\item{select}{A vector of column names}
}
\value{
A string with the constructed SQL query from the provided query
parameter.
}
\description{
Make an SQL query from a list of query parameters
}
\keyword{internal}
