% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_servers.R
\name{server_make_sql_expression}
\alias{server_make_sql_expression}
\title{Convert R expression into SQL expression}
\usage{
server_make_sql_expression(filter, equivalence_table)
}
\arguments{
\item{filter}{a string with the R expression}

\item{equivalence_table}{a data frame with the common R logical operators
and their correspondent SQL operators}
}
\value{
a string of SQL expression
}
\description{
Convert R expression into SQL expression
}
\keyword{internal}
