% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rdbms-helpers.R
\name{server_make_sql_expression}
\alias{server_make_sql_expression}
\title{Convert R expression into SQL expression}
\usage{
server_make_sql_expression(filter)
}
\arguments{
\item{filter}{A string with the R expression}
}
\value{
A string of SQL expression
}
\description{
Convert R expression into SQL expression
}
\keyword{internal}
