% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_servers.R
\name{server_make_subsetting_query}
\alias{server_make_subsetting_query}
\title{Create a subsetting query}
\usage{
server_make_subsetting_query(filter, target_columns, equivalence_table, table)
}
\arguments{
\item{filter}{an expression that will be used to subset on rows}

\item{target_columns}{a comma-separated list of column names to be returned}

\item{equivalence_table}{a table with the common R operators and their
equivalent SQL operator}

\item{table}{the name of the table of interest}
}
\value{
a string with the SQL query made from the input arguments
}
\description{
Create a subsetting query
}
\keyword{internal}
