% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rdbms-helpers.R
\name{server_make_subsetting_query}
\alias{server_make_subsetting_query}
\title{Create a subsetting query}
\usage{
server_make_subsetting_query(filter, target_columns, table)
}
\arguments{
\item{filter}{An expression that will be used to subset on rows}

\item{target_columns}{A comma-separated list of column names to be returned}

\item{table}{The name of the table of interest}
}
\value{
A string with the SQL query made from the input arguments
}
\description{
Create a subsetting query
}
\keyword{internal}
