% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stuff.R
\name{show_example_file}
\alias{show_example_file}
\title{function to display the structure of the credentials file}
\usage{
show_example_file()
}
\description{
function to display the structure of the credentials file
}
\examples{
show_example_file()
}
