% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tables.R
\name{show_tables}
\alias{show_tables}
\title{Display the list of tables in a database}
\usage{
show_tables(credentials.file, project.id, driver.name)
}
\arguments{
\item{credentials.file}{the path to the file with the user-specific credential details for the projects of interest. See the help of the \code{readepi} function for more details.}

\item{project.id}{the name of the target database}

\item{driver.name}{the name of the MS driver. use \code{odbc::odbcListDrivers()} to display the list of installed drivers}
}
\value{
the list of tables in the specified database
}
\description{
Display the list of tables in a database
}
\examples{
\dontrun{
show_tables(
  credentials.file = system.file("extdata", "fake_test.ini", package = "readepi"),
  project.id = "IBS_BHDSS",
  driver.name = "ODBC Driver 17 for SQL Server"
)
}
}
