% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tables.R
\name{show_tables}
\alias{show_tables}
\title{Display the list of tables in a database}
\usage{
show_tables(login)
}
\arguments{
\item{login}{The connection object obtained from the \code{login()} function.}
}
\value{
a \code{character} that contains the list of all tables found
in the specified database.
}
\description{
Display the list of tables in a database
}
\examples{
\dontrun{
# connect to the test MySQL server
  rdbms_login <- login(
    from        = "mysql-rfam-public.ebi.ac.uk",
    type        = "mysql",
    user_name   = "rfamro",
    password    = "",
    driver_name = "",
    db_name     = "Rfam",
    port        = 4497
  )

# display the list of available tables from this database
tables <- show_tables(login = rdbms_login)
}
}
