% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sormas-helpers.R
\name{sormas_get_diseases}
\alias{sormas_get_diseases}
\title{Get the list of disease names from a SORMAS instance}
\usage{
sormas_get_diseases(base_url, user_name, password)
}
\arguments{
\item{base_url}{A character with the base URL to SORMAS system of interest}

\item{user_name}{A character with the user name}

\item{password}{A character with the user's password}
}
\value{
A vector of the list of disease names in a SORMAS instance
}
\description{
Get the list of disease names from a SORMAS instance
}
\examples{
disease_names <- sormas_get_diseases(
  base_url = "https://demo.sormas.org/sormas-rest",
  user_name = "SurvSup",
  password = "Lk5R7JXeZSEc"
)

}
