% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sormas-helpers.R
\name{sormas_get_pathogen_data}
\alias{sormas_get_pathogen_data}
\title{Get pathogen tests data from SORMAS}
\usage{
sormas_get_pathogen_data(login, since)
}
\arguments{
\item{login}{A list with the user's authentication details}

\item{since}{A Date value in ISO8601 format (YYYY-mm-dd). Default is \code{0} i.e.
to fetch all cases from the beginning of data collection.}
}
\value{
A data frame with the following two columns: 'case_id', 'Ct_values'
}
\description{
Get pathogen tests data from SORMAS
}
\keyword{internal}
