% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_select_data}
\alias{sql_select_data}
\title{Subset data read from servers}
\usage{
sql_select_data(
  table_names,
  dbms,
  id_col_name,
  fields,
  records,
  id_position,
  driver_name,
  host,
  database_name,
  user,
  password,
  port
)
}
\arguments{
\item{table_names}{the name of the tables where the data was fetched from}

\item{dbms}{the database management system type}

\item{id_col_name}{the column names that unique identify the records in the
tables}

\item{fields}{a vector of strings where each string is a comma-separated list
of column names.}

\item{records}{a vector or a comma-separated string of subset of subject IDs.}

\item{id_position}{a vector of the column positions of the variable that
unique identifies the subjects in each table}

\item{driver_name}{the driver name}

\item{host}{host server name}

\item{database_name}{the database name}

\item{user}{the user name}

\item{password}{the user's password}

\item{port}{the server port ID}
}
\value{
a \code{list} of 1 or more elements of type \code{data.frame} where every
element contains the subset of the data from the corresponding table
}
\description{
Subset data read from servers
}
\examples{
\dontrun{
  result = sql_select_data(
    table_names = "author",
    dbms = "MySQL",
    id_col_name = "author_id",
    fields = c("author_id", "name"),
    records = NULL,
    id_position = NULL,
    driver_name = "",
    host = "mysql-rfam-public.ebi.ac.uk",
    database_name = "Rfam",
    user = "rfamro",
    password = "",
    port = 4497
  )
}

}
