% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_select_entire_dataset}
\alias{sql_select_entire_dataset}
\title{Fetch entire dataset in a table}
\usage{
sql_select_entire_dataset(table, con)
}
\arguments{
\item{table}{the table name}

\item{con}{the connection object}
}
\value{
a data frame with the entire dataset that is contained in the table
}
\description{
Fetch entire dataset in a table
}
\examples{
result <- sql_select_entire_dataset(
 table = "author",
 con = pool::dbPool(RMySQL::MySQL(),
   driver = "",
   host = "mysql-rfam-public.ebi.ac.uk",
   dbname = "Rfam",
   user = "rfamro",
   password = "",
   port = 4497
 )
)

}
