% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_select_entire_dataset}
\alias{sql_select_entire_dataset}
\title{Fetch entire dataset in a table}
\usage{
sql_select_entire_dataset(
  table,
  dbms,
  driver_name,
  host,
  database_name,
  user,
  password,
  port
)
}
\arguments{
\item{table}{the table name}

\item{dbms}{the database management system type}

\item{driver_name}{the driver name}

\item{host}{host server name}

\item{database_name}{the database name}

\item{user}{the user name}

\item{password}{the user's password}

\item{port}{the server port ID}
}
\value{
an object of type \code{data.frame} with the entire dataset fetched from
the specified table
}
\description{
Fetch entire dataset in a table
}
\examples{
\dontrun{
result <- sql_select_entire_dataset(
  table = "author",
  dbms = "MySQL",
  driver_name = "",
  host = "mysql-rfam-public.ebi.ac.uk",
  database_name = "Rfam",
  user = "rfamro",
  password = "",
  port = 4497
)
}

}
