% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_select_fields_only}
\alias{sql_select_fields_only}
\title{Select specified fields from a table}
\usage{
sql_select_fields_only(
  table,
  field,
  dbms,
  driver_name,
  host,
  database_name,
  user,
  password,
  port
)
}
\arguments{
\item{table}{the table name}

\item{field}{a vector of column names of interest}

\item{dbms}{the database management system type}

\item{driver_name}{the driver name}

\item{host}{host server name}

\item{database_name}{the database name}

\item{user}{the user name}

\item{password}{the user's password}

\item{port}{the server port ID}
}
\value{
a data frame with the specified fields
}
\description{
Select specified fields from a table
}
\examples{
\dontrun{
result <- sql_select_fields_only(
 table = "author",
 field = c("author_id", "name", "last_name"),
 dbms = "MySQL",
 driver_name = "",
 host = "mysql-rfam-public.ebi.ac.uk",
 database_name = "Rfam",
 user = "rfamro",
 password = "",
 port = 4497
)
}

}
