% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_select_fields_only}
\alias{sql_select_fields_only}
\title{Select specified fields from a table}
\usage{
sql_select_fields_only(table, field, con)
}
\arguments{
\item{table}{the table name}

\item{field}{a vector of column names of interest}

\item{con}{the connection object}
}
\value{
a data frame with the specified fileds
}
\description{
Select specified fields from a table
}
\examples{
result <- sql_select_fields_only(
 table = "author",
 field = c("author_id", "name", "last_name"),
 con = DBI::dbConnect(RMySQL::MySQL(),
   driver = "",
   host = "mysql-rfam-public.ebi.ac.uk",
   dbname = "Rfam",
   user = "rfamro",
   password = "",
   port = 4497
 )
)

}
