% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_select_records_and_fields}
\alias{sql_select_records_and_fields}
\title{Select specified records and fields from a table}
\usage{
sql_select_records_and_fields(
  table,
  record,
  con,
  id_column_name,
  field,
  dbms,
  id_pos
)
}
\arguments{
\item{table}{the table name}

\item{record}{a vector or a comma-separated string of subset of subject IDs.}

\item{con}{the connection object}

\item{id_column_name}{the column names that unique identify the records in
the tables}

\item{field}{a vector of strings where each string is a comma-separated list
of column names.}

\item{dbms}{the database management system type}

\item{id_pos}{a vector of the column positions of the variable that
unique identifies the subjects in each table}
}
\value{
a data frame with the specified columns and records
}
\description{
Select specified records and fields from a table
}
\examples{
result <- sql_select_records_and_fields(
table = "author",
record = c("1", "20", "50"),
con = DBI::dbConnect(RMySQL::MySQL(),
driver = "",
host = "mysql-rfam-public.ebi.ac.uk",
dbname = "Rfam",
user = "rfamro",
password = "",
port = 4497),
id_column_name = "author_id",
field = c("author_id", "last_name"),
dbms = "MySQL",
id_pos = NULL)

}
