% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{sql_select_records_only}
\alias{sql_select_records_only}
\title{Select specified records from a table}
\usage{
sql_select_records_only(table, record, con, dbms, id_column_name, id_pos)
}
\arguments{
\item{table}{the table name}

\item{record}{a vector or a comma-separated string of subset of subject IDs.}

\item{con}{the connection object}

\item{dbms}{the database management system type}

\item{id_column_name}{the column names that unique identify the records in
the tables}

\item{id_pos}{a vector of the column positions of the variable that
unique identifies the subjects in each table}
}
\value{
a data frame with the records of interest
}
\description{
Select specified records from a table
}
\examples{
result <- sql_select_records_only(
 table = "author",
 record = c("1", "20", "50"),
 con = pool::dbPool(RMySQL::MySQL(),
   driver = "",
   host = "mysql-rfam-public.ebi.ac.uk",
   dbname = "Rfam",
   user = "rfamro",
   password = "",
   port = 4497
 ),
 dbms = "MySQL",
 id_column_name = NULL,
 id_pos = 1
)
}
