% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers.R
\name{sql_server_read_data}
\alias{sql_server_read_data}
\title{Read data from relational databases hosted by a MS SQL server.}
\usage{
sql_server_read_data(
  user,
  password,
  host,
  port = 1433,
  database_name,
  driver_name,
  source = NULL,
  records = NULL,
  fields = NULL,
  id_position = NULL,
  id_col_name = NULL,
  dbms
)
}
\arguments{
\item{user}{the user name}

\item{password}{the user password}

\item{host}{the name of the host server}

\item{port}{the port ID}

\item{database_name}{the name of the database that contains the table from
which the data should be pulled}

\item{driver_name}{the name of the MS driver. use \code{odbc::odbcListDrivers()}
to display the list of installed drivers}

\item{source}{and SQL query or a vector of table names from
the project or database. When this is not specified, the function will
extract data from all tables in the database.}

\item{records}{a vector or a comma-separated string of subset of subject IDs.
When specified, only the records that correspond to these subjects will be
imported.}

\item{fields}{a vector of strings where each string is a comma-separated list
of column names. The element of this vector should be a list of column names
from the first table specified in the \code{table_names} argument and so on...}

\item{id_position}{a vector of the column positions of the variable that
unique identifies the subjects in each table. When the column name with the
subject IDs is known, use the \code{id_col_name} argument instead. default is.
default is 1.}

\item{id_col_name}{the column name with the subject IDs}

\item{dbms}{the SQL server type}
}
\value{
a list of data frames
}
\description{
For a user with read access to a Microsoft SQL server,
this function allows data import from the database into R. It required the
installation
of the appropriate MS driver that is compatible with the SQL server version
hosting the
database.
}
\examples{
\dontrun{
data <- sql_server_read_data(
  user = "rfamro",
  password = "",
  host = "mysql-rfam-public.ebi.ac.uk",
  port = 4497,
  database_name = "Rfam",
  source = "author",
  driver_name = "",
  dbms = "MySQL"
)
}
}
