% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readepi_helpers.R
\name{subset_fields}
\alias{subset_fields}
\title{Subset fields}
\usage{
subset_fields(data.frame, fields, table.name)
}
\arguments{
\item{data.frame}{the input data frame}

\item{fields}{the list of columns of interest}

\item{table.name}{the table names}
}
\value{
a list of 2 elements: the subset data frame and an integer that tells whether all fields were missing in the table (1) or not (0)
}
\description{
Subset fields
}
\examples{
\dontrun{
data <- subset_fields(
  data.frame = data.frame,
  fields = "date,sex,age",
  table.name = "covid"
)
}
}
