% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rdbms-helpers.R
\name{url_check}
\alias{url_check}
\title{Check if the value for the base_url argument has a correct structure}
\usage{
url_check(base_url)
}
\value{
throws an error if the domain of the provided URL is not valid,
(invisibly) TRUE
}
\description{
Check if the value for the base_url argument has a correct structure
}
\keyword{internal}
