% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_servers_helpers.R
\name{visualise_table}
\alias{visualise_table}
\title{Visualize the first 5 rows of the data from a table}
\usage{
visualise_table(table, con, dbms, display = TRUE)
}
\arguments{
\item{table}{the table name}

\item{con}{the connection object}

\item{dbms}{the database management system type}

\item{display}{a boolean that specify whether to display the first 5 rows or
not. default = TRUE}
}
\value{
return the first 5 rows of the table if display=TRUE
}
\description{
Visualize the first 5 rows of the data from a table
}
\examples{
visualise_table(
 table = "author",
 con = DBI::dbConnect(RMySQL::MySQL(),
   driver = "",
   host = "mysql-rfam-public.ebi.ac.uk",
   dbname = "Rfam",
   user = "rfamro",
   password = "",
   port = 4497
 ),
   display = TRUE,
   dbms =  "MySQL"
)

}
